\name{aflatoxin}
\alias{aflatoxin}
\docType{data}
\title{aflatoxin dosage and liver cancer in lab animals}
\description{
  Aflatoxin B1 was fed to lab animals at vary doses and
  the number responding with liver cancer recorded.
}
\usage{data(aflatoxin)}
\format{
  A data frame with 6 observations on the following 3 variables.
  \describe{
    \item{dose}{dose in ppb}
    \item{total}{number of test animals}
    \item{tumor}{number with liver cancer}
  }
}
\source{
  Gaylor DW (1987) "Linear nonparametric upper limits for low dose
  extrapolation" ASA Proceedings of the Biopharmaceutical Section.
}
\examples{
data(aflatoxin)
}
\keyword{datasets}
