% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{africa}
\alias{africa}
\title{miltary coups and politics in sub-Saharan Africa}
\format{
A data frame with 47 observations on the following 9 variables.
\describe{ \item{miltcoup}{number of successful military coups from
independence to 1989} \item{oligarchy}{number years country ruled by
military oligarchy from independence to 1989} \item{pollib}{Political
liberalization - 0 = no civil rights for political expression, 1 = limited
civil rights for expression but right to form political parties, 2 = full
civil rights} \item{parties}{Number of legal political parties in 1993}
\item{pctvote}{Percent voting in last election} \item{popn}{Population in
millions in 1989} \item{size}{Area in 1000 square km} \item{numelec}{Total
number of legislative and presidential elections} \item{numregim}{Number of
regime types} }
}
\source{
Bratton, Michael, and Nicholas Van De Walle.  1997.  ``Political
Regimes and Regime Transitions in Africa, 1910-1994.'' \emph{Study Number
I06996.} Ann Arbor: Inter-University Consortium for Political and Social
Research.
}
\description{
Data is a subset of a larger study on factors affecting regime stability in
Sub-Saharan Africa
}
\references{
"Bayesian Methods: A Social and Behavioral Sciences Approach" by
Jeff Gill 2002.
}
\keyword{datasets}
