% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{alfalfa}
\alias{alfalfa}
\title{Effects of seed inoculum, irrigation and shade on alfalfa yield}
\format{
This data frame contains the following columns: \describe{
\item{shade}{ Distance of location from tree line divided into 5
shade areas } \item{irrigation}{ Irrigation effect divided into 5
levels } \item{inoculum}{ Four types of seed incolum, A-D with E as
control. } \item{yield}{ Dry matter yield of alfalfa }}
}
\source{
Petersen, R.G. 1994. Agricultural Field Experiments, Design and
Analysis. Marcel Dekker, Inc., New York.  Pages 70-74. 1994
}
\description{
The \code{alfalfa} data frame has 25 rows and 4 columns. Data comes from an
experiment to test the effects of seed inoculum, irrigation and shade on
alfalfa yield. A latin square design has been used.
}
\keyword{datasets}
