\name{beetle}
\alias{beetle}
\docType{data}
\title{
Beetles exposed to fumigant
}
\description{
Grain beetles were exposed to ethylene oxide
}
\usage{data(beetle)}
\format{
  A data frame with 10 observations on the following 3 variables.
  \describe{
    \item{\code{conc}}{concentration of ethylene oxide in mg/l}
    \item{\code{affected}}{number affected}
    \item{\code{exposed}}{number exposed}
  }
}
\source{
Busvine (1938)
}
\references{
Collet D. "Modelling Binary Data"
}
\examples{
data(beetle)
## maybe str(beetle) ; plot(beetle) ...
}
\keyword{datasets}
