\name{breaking}
\alias{breaking}
\title{Breaking strengths of material by day, supplier, operator}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(breaking)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
