\name{cathedral}
\alias{cathedral}
\title{Cathedral nave heights and lengths in England}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(cathedral)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
