\name{chicago}
\alias{chicago}
\title{Chicago insurance redlining}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(chicago)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
