\name{chiczip}
\alias{chiczip}
\title{Chicago zip codes north-south}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(chiczip)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
