\name{chmiss}
\alias{chmiss}
\title{Chicago insurance redlining}
\description{Data from a 1970's study on the relationship between
  insurance redlining in Chicago and racial composition, fire and
  theft rates, age of housing and income in 47 zip codes. Missing
  values have been randomly added.
}
\usage{data(chmiss)
}
\format{This dataframe contains the following columns 
  \describe{

    \item{\code{race}}{
            racial composition in percent minority 
      }
\item{\code{fire}}{
      fires per 100 housing units }
    \item{\code{theft}}{
      theft per 1000 population }
\item{\code{age}}{
      percent of housing units built before 1939 }
\item{\code{involact}}{ 
      new FAIR plan policies and renewals per 100 housing units }
\item{\code{income}}{
      median family income in thousands of dollars}
    \item{\code{side}}{ North or South side of Chicago}
      
  }
}
\source{Adapted from  "Data : A Collection of Problems from Many Fields
  for the Student and Research Worker"  by D. Andrews and A. Herzberg
  published by Springer-Verlag, in 1985}
}
\keyword{datasets}
