% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{choccake}
\alias{choccake}
\title{Chocolate cake experiment with split plot design}
\format{
A data frame with 270 observations on the following 4 variables.
\describe{ \item{recipe}{Chocolate for recipe 1 was added at 40C,
Chocolate for recipe 2 was added at 60C and recipe 3 had extra sugar}
\item{batch}{batch number from 1 to 15}
\item{temp}{temperature at which cake was baked: \code{175C}
\code{185C} \code{195C} \code{205C} \code{215C} \code{225C}}
\item{breakang}{the breaking angle of the cake} }
}
\source{
Cochran W. and Cox G. (1992) Experimental Designs, 2nd Edition Wiley
}
\description{
An experiment was conducted to determine the effect of recipe and baking
temperature on chocolate cake quality. 15 batches of cake mix for each
recipe were prepared. Each batch was sufficient for six cakes. Each of the
six cakes was baked at a different temperature which was randomly assigned.
Several measures of cake quality were recorded of which breaking angle was
just one.
}
\keyword{datasets}
