\name{clot}
\alias{clot}
\title{Blood clotting times}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(clot)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
