% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{cmob}
\alias{cmob}
\title{Social class mobility from 1971 to 1981 in the UK}
\format{
A data frame with 36 observations on the following 3 variables.
\describe{ \item{y}{Frequency of observation} \item{class71}{social class in
1971 - a factor with levels \code{I}, professionals, \code{II}
semi-professionals, \code{IIIN} skilled non-manual, \code{IIIM} skilled
manual, \code{IV} semi-skilled, \code{V} unskilled} \item{class81}{social
class in 1971 - a factor with levels \code{I} \code{II} \code{IIIN}
\code{IIIM} \code{IV} \code{V} with same classification} }
}
\source{
D. Blane and S. Harding and M. Rosato (1999) "Does social mobility
affect the size of the socioeconomic mortality differential?: Evidence from
the Office for National Statistics Longitudinal Study" JRSS-A, 162 59-70.
}
\description{
Social class mobility from 1971 to 1981 for 42425 men from the United
Kingdom census. Subjects were aged 45-64.
}
\keyword{datasets}
