% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{cns}
\alias{cns}
\title{Malformations of the central nervous system}
\format{
A data frame with 16 observations on the following 7 variables.
\describe{ \item{Area}{a factor with levels \code{Cardiff} \code{GlamorganC}
\code{GlamorganE} \code{GlamorganW} \code{MonmouthOther} \code{MonmouthV}
\code{Newport} \code{Swansea} being areas of South Wales} \item{NoCNS}{count
of births with no CNS problem} \item{An}{count of Anencephalus births}
\item{Sp}{count of Spina Bifida births} \item{Other}{count of other CNS
births} \item{Water}{water hardeness} \item{Work}{a factor with levels
\code{Manual} \code{NonManual} being the type of work done by the parents} }
}
\source{
C. Lowe and C. Roberts and S. Lloyd, (1971) Malformations of the
central nervous system and softness of local water supplies, British Medical
Journal, 15,357-361.
}
\description{
Frequencies of various malformations of the central nervous system recorded
on live births in South Wales, UK. Study was designed to determine the
effect of water hardness on the incidence of such malformations.
}
\references{
P. McCullagh and J. Nelder (1989), Generalized Linear Models,
Chapman and Hall, 2nd Ed.
}
\keyword{datasets}
