\name{coagulation}
\alias{coagulation}
\title{Blood coagulation times by diet}
\description{Dataset comes from a study of  blood coagulation
times. 24 animals were randomly assigned to four different diets and
the samples were taken in a random order.
}
\usage{data(coagulation)
}
\format{This dataframe contains the following columns 
  \describe{

    \item{\code{coag}}{
      coagulation time in seconds
    }

    \item{\code{diet}}{
      diet type - A,B,C or D
    }
  }
}

\source{"Statistics for Experimenters" by G. P. Box, W. G. Hunter and
  J. S. Hunter, Wiley, 1978
}
\keyword{datasets}
