\name{composite}
\alias{composite}
\title{
Strength of a thermoplastic composite depending on two factors
}
\description{
The \code{composite} data frame has 9 rows and 3 columns.
Data comes from an experiment to test the strength
of a thermoplastic composite depending on the power of a laser
and speed of a tape.
}
\usage{
data(composite)
}
\format{
This data frame contains the following columns:
\describe{
\item{\code{strength}}{
interply bond strength of the composite
}
\item{\code{laser}}{
laser power at 40, 50 or 60W
}
\item{\code{tape}}{
tape speed, slow=6.42 m/s, medium=13m/s and fast=27m/s
}}}
\source{
  Mazumdar, S and Hoa S (1995) "Application of a Taguchi Method for
  Process enhancement of an online consolidation technique" Composites
  26, 669-673
}
\keyword{datasets}

