\name{cornnit}
\alias{cornnit}
\docType{data}
\title{Corn yields from nitrogen application}
\description{
  The relationship between corn yield
  (bushels per acre) and nitrogen (pounds per acre) fertilizer
  application were studied in Wisconsin.
}
\usage{data(cornnit)}
\format{
  A data frame with 44 observations on the following 2 variables.
  \describe{
    \item{yield}{corn yield in bushels per acre}
    \item{nitrogen}{pounds per acre}
  }
}
\source{
  Unknown
}
\keyword{datasets}
