\name{corrosion}
\alias{corrosion}
\title{Corrosion loss in Cu-Ni alloys}
\description{
  Data consist of thirteen specimens of 90/10 Cu-Ni
alloys with varying iron content in percent. The specimens were
submerged in sea water for 60 days and the weight loss due to corrosion
was recorded in units of milligrams per square decimeter per day.

}
\usage{data(corrosion)
}
\format{This dataframe contains the following columns
  \describe{

    \item{\code{Fe}}{
      Iron content in percent
    }
    \item{\code{loss}}{
      Weight loss in mg per square decimeter per day
    }
  }
}
\source{"Applied Regression Analysis" by N. Draper and H. Smith,
Wiley, 1998
}
\keyword{datasets}
