\name{Cpplot}
\alias{Cpplot}
\title{Cp plot}
\description{Makes a Cp plot
}
\usage{
Cpplot(cp)
}
\arguments{

  \item{cp}{A leaps object returned from leaps()}

}
\value{none
}
\details{
  Requires leaps package
}
\references{
}
\seealso{leaps()
}
\examples{
}
\author{Julian Faraway}
\keyword{regression}
