\name{death}
\alias{death}
\docType{data}
\title{Death penalty in Florida 1977}
\description{
  Data on 326 defendents in homicide indictments in 20 Florida counties
  during 1976-77.
}
\usage{data(death)}
\format{
  A data frame with 8 observations on the following 4 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{penalty}}{Did the subject recieve the death penalty?
      \code{no}  or \code{yes}}
    \item{\code{victim}}{Was the victim \code{b}lack or \code{w}hite?}
    \item{\code{defend}}{Was the defendent \code{b}lack or \code{w}hite?}
  }
}
\source{
  Radelet M. (1981) Racial characteristics and the imposition of the
  death penalty. Amer. Sociol. Rev. \bold{46} 918-927.
}
\references{
  Agresti A. (1990) Categorical Data Analysis, Wiley.
}
\keyword{datasets}
