% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{debt}
\alias{debt}
\title{psychology of debt}
\format{
A data frame with 464 observations on the following 13 variables.
\describe{ \item{incomegp}{income group (1=lowest, 5=highest)}
\item{house}{security of housing tenure (1=rent, 2=mortgage, 3=owned
outright)} \item{children}{number of children in household}
\item{singpar}{is the respondent a single parent?} \item{agegp}{age group
(1=youngest)} \item{bankacc}{does the respondent have a bank account?}
\item{bsocacc}{does the respondent have a building society account?}
\item{manage}{self-rating of money management skill (high values=high
skill)} \item{ccarduse}{how often did s/he use credit cards (1=never...
3=regularly)} \item{cigbuy}{does s/he buy cigarettes?} \item{xmasbuy}{does
s/he buy Christmas presents for children?} \item{locintrn}{score on a locus
of control scale (high values=internal)} \item{prodebt}{score on a scale of
attitudes to debt (high values=favourable to debt} }
}
\source{
Lea, Webley & Walker, 1995, Journal of Economic Psychology, 16,
181-201 Data obtained from \url{http://www.exeter.ac.uk/~SEGLea}.
}
\description{
The data arise from a large postal survey on the psychology of debt.
}
\details{
All yes/no questions are coded 0=no, 1=yes. Locus of control is a
personality measure introduced by Rotter, which claims to differentiate
people according to how much they feel things that happen to them are as a
result of processes within themselves (internal locus of control) or outside
events (external locus of control).
}
\keyword{datasets}
