\name{diabetes}
\alias{diabetes}
\docType{data}
\title{Diabetes and obesity, cardiovascular risk factors}
\description{
403 African Americans were interviewed in a study to understand the
prevalence of obesity, diabetes, and other cardiovascular risk factors
in central Virginia.
}
\usage{data(diabetes)}
\format{
  A data frame with 403 observations on the following 19 variables.
  \describe{
    \item{\code{id}}{Subject ID}
    \item{\code{chol}}{Total Cholesterol}
    \item{\code{stab.glu}}{Stabilized Glucose}
    \item{\code{hdl}}{High Density Lipoprotein}
    \item{\code{ratio}}{Cholesterol/HDL Ratio}
    \item{\code{glyhb}}{Glycosolated Hemoglobin}
    \item{\code{location}}{County - a factor with levels \code{Buckingham} \code{Louisa}}
    \item{\code{age}}{age in years}
    \item{\code{gender}}{a factor with levels \code{male} \code{female}}
    \item{\code{height}}{height in inches}
    \item{\code{weight}}{weight in pounds}
    \item{\code{frame}}{a factor with levels \code{small} \code{medium} \code{large}}
    \item{\code{bp.1s}}{First Systolic Blood Pressure}
    \item{\code{bp.1d}}{First Diastolic Blood Pressure}
    \item{\code{bp.2s}}{Second Systolic Blood Pressure}
    \item{\code{bp.2d}}{Second Diastolic Blood Pressure}
    \item{\code{waist}}{waist in inches}
    \item{\code{hip}}{hip in inches}
    \item{\code{time.ppn}}{Postprandial Time (in minutes) when Labs were Drawn}
  }
}
\details{
Glycosolated hemoglobin greater than 7.0 is usually taken as a positive
diagnosis of diabetes
}
\source{
Willems JP, Saunders JT, DE Hunt, JB
Schorling: Prevalence of coronary heart disease risk factors among
rural blacks: A community-based study. Southern Medical Journal
90:814-820; 1997
}
\references{
Schorling JB, Roach J, Siegel M, Baturka N, Hunt
DE, Guterbock TM, Stewart HL: A trial of church-based smoking
cessation interventions for rural African Americans. Preventive
Medicine 26:92-101; 1997
}
\keyword{datasets}
