% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{dicentric}
\alias{dicentric}
\title{Radiation dose effects on chromosomal abnormality}
\format{
A data frame with 27 observations on the following 4 variables.
\describe{ \item{cells}{Number of cells in hundreds}
\item{ca}{Number of chromosomal abnormalities}
\item{doseamt}{amount of dose in Grays} \item{doserate}{rate
of dose in Grays/hour} }
}
\source{
Purott R. and Reeder E. (1976) The effect of changes in dose rate on
the yield of chromosome aberrations in human lymphocytes exposed to gamma
radiation. Mutation Research. 35, 437-444.
}
\description{
An experiment was conducted to determine the effect of gamma radiation on
the numbers of chromosomal abnormalities observed
}
\references{
Frome E. and DuFrain R. (1986) Maximum Likelihood Estimation for
Cytogenic Dose-Response Curves. Biometrics. 42, 73-84.
}
\keyword{datasets}
