% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{divusa}
\alias{divusa}
\title{Divorce in the USA 1920-1996}
\format{
A data frame with 77 observations on the following 7 variables.
\describe{ \item{year}{the year from 1920-1996} \item{divorce}{divorce per
1000 women aged 15 or more} \item{unemployed}{unemployment rate}
\item{femlab}{percent female participation in labor force aged 16+}
\item{marriage}{marriages per 1000 unmarried women aged 16+}
\item{birth}{births per 1000 women aged 15-44} \item{military}{military
personnel per 1000 population} }
}
\source{
Unknown
}
\description{
Divorce rates in the USA from 1920-1996
}
\keyword{datasets}
