\name{drugpsy}
\alias{drugpsy}
\title{Drug treatment of psychiatric patients}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(drugpsy)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
