\name{eco}
\alias{eco}
\title{Ecological regression example}
\description{Relationship
between 1998 per capita income dollars from all sources and the proportion
of legal state residents born in the United States in 1990 for each of
the 50 states plus the District of Columbia
}
\usage{data(eco)
}
\format{This dataframe contains the following columns
  \describe{
    \item{\code{usborn}}{
      Percentage of population born in the United States}
    \item{\code{income}}{
      Per capita annual income in dollars}
    \item{\code{home}}{
      Percentage born in state}
    \item{\code{pop}}{
      Population of state
    }
  }
}
\source{US Bureau of the Census}
\keyword{datasets}
