\name{eggs}
\alias{eggs}
\title{Laboratory testing of dried egg fat content}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(eggs)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
