\name{esdcomp}
\alias{esdcomp}
\docType{data}
\title{Complaints about emergency room doctors}
\description{
  Data was recorded on 44 doctors working in an emergency service
  at a hospital to study the factors affecting the number of complaints
  received.
}
\usage{data(esdcomp)}
\format{
  A data frame with 44 observations on the following 6 variables.
  \describe{
    \item{\code{visits}}{the number of patient visits}
    \item{\code{complaints}}{the number of complaints}
    \item{\code{residency}}{is the doctor in residency training \code{N}
      or \code{Y}}
    \item{\code{gender}}{gender of doctor \code{F}  or \code{M}}
    \item{\code{revenue}}{dollars per hour earned by the doctor}
    \item{\code{hours}}{total number of hours worked}
  }
}
\source{
  Chap T. Le (1998) "Applied Categorical Data Analysis" Wiley
}
\keyword{datasets}
