\name{exa}
\alias{exa}
\title{Non parametric regression test data A}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(exa)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
