\name{exb}
\alias{exb}
\title{Non parametric regression test data B}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(exb)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
