\name{faithful}
\alias{faithful}
\title{Eruption durations and intervals for Old Faithful}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(faithful)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
