% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{femsmoke}
\alias{femsmoke}
\title{Mortality due to smoking according age group in women}
\format{
A data frame with 28 observations on the following 4 variables.
\describe{ \item{y}{observed count for given combination} \item{smoker}{a
factor with levels \code{yes} \code{no}} \item{dead}{a factor with levels
\code{yes} \code{no}} \item{age}{a factor with agegroup levels \code{18-24}
\code{25-34} \code{35-44} \code{45-54} \code{55-64} \code{65-74} \code{75+}}
}
}
\source{
D. Appleton, J. French, M. Vanderpump (1996) "Ignoring a Covariate:
An Example of Simpson's Paradox" American Statistician, 50, 340-341
}
\description{
In 1972-74, a survey of one in six residents of Whickham, near Newcastle,
England was made. Twenty years later, this data recorded in a follow-up
study. Only women who are current smokers or who have never smoked are
included.
}
\keyword{datasets}
