\name{crap}
\alias{crap}
\title{Insect damage to wheat varieties}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(crap)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
