% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{fruitfly}
\alias{fruitfly}
\title{Longevity of fruitflies depending on sexual activity and thorax length}
\format{
This data frame contains the following columns: \describe{

\item{thorax}{ Thorax length } \item{longevity}{ Lifetime in
days } \item{activity}{ The group: isolated = fly kept solitary, one
= fly kept with one pregnant fruitfly, many = fly kept with eight pregnant
fruitflies, low= fly kept with one virgin fruitfly, high = fly kept with
eight virgin fruitflies.  }}
}
\source{
"Sexual Activity and the Lifespan of Male Fruitflies" by L.
Partridge and M. Farquhar, Nature, 1981, 580-581
}
\description{
The \code{fruitfly} data frame has 9 rows and 3 columns.  125 fruitflies
were divided randomly into 5 groups of 25 each. The response was the
longevity of the fruitfly in days. One group was kept solitary, while
another was kept individually with a virgin female each day. Another group
was given 8 virgin females per day. As an additional control the fourth and
fifth groups were kept with one or eight pregnant females per day.  Pregnant
fruitflies will not mate. The thorax length of each male was measured as
this was known to affect longevity. One observation in the many group has
been lost.
}
\keyword{datasets}
