\name{haireye}
\alias{haireye}
\docType{data}
\title{Hair and eye color}
\description{
  Data collected from 592 students in an introductory statistics class
}
\usage{data(haireye)}
\format{
  A data frame with 16 observations on the following 3 variables.
  \describe{
    \item{y}{count of the number of student with given hair/eye combination}
    \item{eye}{a factor with levels \code{green} \code{hazel} \code{blue} \code{brown}}
    \item{hair}{a factor with levels \code{BLACK} \code{BROWN} \code{RED} \code{BLOND}}
  }
}
\source{
  Snee R. (1974) Graphical display of two-way contingency
  tables. American Statistician, 28, 9-12
}
\keyword{datasets}
