\name{hormone}
\alias{hormone}
\title{Hormones and Sexual Orientation}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(hormone)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
