\name{hsb}
\alias{hsb}
\docType{data}
\title{Career choice of high school students}
\description{
  Data was collected as a subset
of the "High School and Beyond" study conducted by the
National Education Longitudinal Studies (NELS) program
of the National Center for Education Statistics (NCES).
}
\usage{data(hsb)}
\format{
  A data frame with 200 observations on the following 11 variables.
  \describe{
    \item{\code{id}}{ID of student}
    \item{\code{gender}}{a factor with levels \code{female} \code{male}}
    \item{\code{race}}{a factor with levels \code{african-amer} \code{asian} \code{hispanic} \code{white}}
    \item{\code{ses}}{socioeconomic class - a factor with levels \code{high} \code{low} \code{middle}}
    \item{\code{schtyp}}{school type - a factor with levels \code{private} \code{public}}
    \item{\code{prog}}{choice of high school program - a factor with levels \code{academic} \code{general} \code{vocation}}
    \item{\code{read}}{reading score}
    \item{\code{write}}{writing score}
    \item{\code{math}}{math score}
    \item{\code{science}}{science score}
    \item{\code{socst}}{social science score}
  }
}
\details{
  One purpose of the study was 
to determine which factors are related to the choice of the
type of program, academic,
vocational or general, that the students pursue in high school.
}
\source{
  National Education Longitudinal Studies (NELS) program
of the National Center for Education Statistics (NCES).
}
\keyword{datasets}
