\name{infmort}
\alias{infmort}
\title{Infant mortality according to income and region}
\description{The \code{infmort} data frame has 105 rows and 4 columns.
  The infant mortality in regions of the world may be related to
  per capita income and whether oil is exported. The dataset is not recent.
}
\usage{
data(infmort)
}
\format{
This data frame contains the following columns:
\describe{

  \item{\code{region}}{
    Region of the world, Africa, Europe, Asia or the Americas
  }
  \item{\code{income}}{
    Per capita annual income in dollars
  }

  \item{\code{mortality}}{
      Infant mortality in deaths per 1000 births
  }
  \item{\code{oil}}{
    Does the country export oil or not?
  }
  }}
\source{Unknown}
\keyword{datasets}
