\name{irrigation}
\alias{irrigation}
\title{Agricultural experiment with irrigation }
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(irrigation)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
