\name{jsp}
\alias{jsp}
\title{Junior Schools Project}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(jsp)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
