\name{kanga}
\alias{kanga}
\title{Historic Kangaroos}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(kanga)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
