\name{leafblotch}
\alias{leafblotch}
\title{Leaf blotch on Barley}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(leafblotch)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
