\name{leafburn}
\alias{leafburn}
\docType{data}
\title{
Data on the burning time of samples of tobacco leaves 
}
\description{
Data on the burning time of samples of tobacco leaves 
}
\usage{data(leafburn)}
\format{
  A data frame with 30 observations on the following 4 variables.
  \describe{
    \item{\code{nitrogen}}{nitrogen content by percentage weight}
    \item{\code{chlorine}}{chlorine content by percentage weight}
    \item{\code{potassium}}{potassium content by percentage weight}
    \item{\code{burntime}}{burn time in seconds}
  }
}
\source{
Steel, R. G. D. and Torrie, J. H. (1980), Principles and Procedures of Statistics, Second Edition, New York: McGraw-Hill
}
\keyword{datasets}
