\name{maxadjr}
\alias{maxadjr}
\title{Maximum Adjusted R-squared}
\description{Displays the best models from a leaps object
}
\usage{
maxadjr(l,best=3)
}
\arguments{

  \item{l}{A leaps object returned from leaps()}

  \item{best}{An optional argument specify the number of models to be
    returned taking the default value of 3}
}
\value{A list of the best models
}
\details{
  Requires leaps package
}
\references{
}
\seealso{leaps()
}
\examples{
}
\author{Julian Faraway}
\keyword{regression}
