\name{melanoma}
\alias{melanoma}
\docType{data}
\title{Melanoma by type and location}
\description{
  Data comes from a study of Malignant Melanoma involving 400 subjects.
}
\usage{data(melanoma)}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{count}}{number of cases}
    \item{\code{tumor}}{type of tumor - a factor with levels \code{freckle} \code{indeterminate} \code{nodular} \code{superficial}}
    \item{\code{site}}{location of tumor on the body - a factor with levels \code{extremity} \code{head} \code{trunk}}
  }
}
\source{
  Dobson A. (2002) An introduction to generalized linear models, Chapman Hall.
}
\keyword{datasets}
