% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{motorins}
\alias{motorins}
\title{Third party motor insurance claims in Sweden in 1977}
\format{
A data frame with 1797 observations on the following 8 variables.
\describe{ \item{Kilometres}{an ordered factor representing kilomoters per
year with levels 1: < 1000, 2: 1000-15000, 3: 15000-20000, 4: 20000-25000,
5: > 25000} \item{Zone}{a factor representing geographical area with levels
1: Stockholm, Goteborg, Malmo with surroundings 2: Other large cities with
surroundings 3: Smaller cities with surroundings in southern Sweden 4: Rural
areas in southern Sweden 5: Smaller cities with surroundings in northern
Sweden 6: Rural areas in northern Sweden 7: Gotland} \item{Bonus}{No claims
bonus. Equal to the number of years, plus one, since last claim}
\item{Make}{A factor representing eight different common car models. All
other models are combined in class 9} \item{Insured}{Number of insured in
policy-years} \item{Claims}{Number of claims} \item{Payment}{Total value of
payments in Skr} \item{perd}{payment per claim} }
}
\source{
\url{http://www.statsci.org/data/general/motorins.html}
}
\description{
In Sweden all motor insurance companies apply identical risk arguments to
classify customers, and thus their portfolios and their claims statistics
can be combined. The data were compiled by a Swedish Committee on the
Analysis of Risk Premium in Motor Insurance. The Committee was asked to look
into the problem of analyzing the real influence on claims of the risk
arguments and to compare this structure with the actual tariff.
}
\references{
Hallin, M., and Ingenbleek, J.-F. (1983). The Swedish automobile
portfolio in 1977. A statistical study. Scandinavian Actuarial Journal,
49-64.
}
\keyword{datasets}
