% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{neighbor}
\alias{neighbor}
\title{Questionnaire study of neighborly help}
\format{
A data frame with 181 observations on the following 8 variables.
\describe{ \item{longlive}{About how long have you lived where you do now?
Ans is a factor with levels \code{<6mos} \code{6-12mos} \code{1-3yrs}
\code{3-10yrs} \code{10yrs}} \item{wherebfr}{Where were you living before
you moved to your present house? Ans is a factor with levels \code{same}
\code{Exeter} \code{Devon} \code{Britain} \code{Abroad}} \item{hownbly}{How
neighborly do you think the area where you now live is? Ans is a factor with
levels \code{Vunfriendly} \code{NVfriendly} \code{Average} \code{FFriendly}
\code{VFriendly}} \item{knowname}{Roughly how many people in your street, or
in the streets just near you, do you know the names of? Ans is a factor with
levels \code{none} \code{1-5} \code{6-20} \code{20+}} \item{callname}{How
many of those people (not counting children) would you call by their first
names? Ans is a factor with levels \code{none} \code{1-5} \code{6-20}
\code{20+}} \item{age}{a factor with levels \code{-18} \code{18-30}
\code{31-50} \code{51-65} \code{65+}} \item{district}{a factor with levels
\code{1} \code{2} \code{3} \code{4}} \item{sex}{a factor with levels
\code{female} \code{male}} }
}
\source{
P. Webley & S. Lea 1993, Human Relations 46, 65-76.
}
\description{
Subjects were asked questions in a study of neighborly help. Questions below
are a subset of the full study.
}
\details{
Exeter is a city in the county of Devon which is in Britain.  The four
districts can be briefly described as follows. District 1 was a
long-established residential area near the city centre, with housing dating
from the late nineteenth century. Originally working class, it now has a
considerable middle class population with some student and other temporary
accommodation. District 2 was a working-class housing estate dating from the
1930s, with mainly rented accommodation but some owner occupation. District
3 was the oldest part of a more recently developed, mainly middle-class,
almost exclusively owner-occupied estate, dating from the 1960s. District 4
was the most recently developed part of a more sought-after middle-class
residential area, with smaller but almost entirely owner-occupied properties
dating from the 1970s and 1980s.
}
\keyword{datasets}
