\name{nels88}
\alias{nels88}
\title{Subset of National Education Longitudinal Study 1988}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(nels88)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
