\name{nepali}
\alias{nepali}
\docType{data}
\title{Nepali child heath study}
\description{
  The data are a subset from public health
study on Nepalese children.
}
\usage{data(nepali)}
\format{
  A data frame with 1000 observations on the following 9 variables.
  \describe{
    \item{\code{id}}{There is a seven digit code for the child's ID:  2 digits for the panchayat
  number; 1 digit for the ward within panchayat; 3 digits for the household;
  1 digit for child within household.}
    \item{\code{sex}}{1 = male; 2 =  female}
    \item{\code{wt}}{Child's weight measured in kilograms}
    \item{\code{ht}}{Child's height measured in centimeters}
    \item{\code{mage}}{Mother's age in years}
    \item{\code{lit}}{Indicator of mother's literacy:  0 = no; 1 = yes}
    \item{\code{died}}{The number of children the mother has had that died.}
    \item{\code{alive}}{The number of children the mother has ever had born alive}
    \item{\code{age}}{age of child}
  }
}
\source{
West KP, Jr., LeClerq SC, Shrestha SR, Wu LS, Pradhan EK, Khatry SK, Katz J,
Adhikari R, Sommer A.  Effects of vitamin A on growth of vitamin A deficient
children: field studies in Nepal.
J Nutr 1997;10:1957-1965.
}
\keyword{datasets}
