% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{nes96}
\alias{nes96}
\title{US 1996 national election study}
\format{
A data frame with 944 observations on the following 10 variables.
\describe{ \item{popul}{population of respondent's location in 1000s of
people} \item{TVnews}{days in the past week spent watching news on TV}
\item{selfLR}{Left-Right self-placement of respondent: an ordered factor
with levels extremely liberal, \code{extLib} < liberal, \code{Lib} <
slightly liberal, \code{sliLib} < moderate, \code{Mod} < slightly
conservative, \code{sliCon} < conservative, \code{Con} < extremely
conservative, \code{extCon}} \item{ClinLR}{Left-Right placement of Bill
Clinton (same scale as selfLR): an ordered factor with levels \code{extLib}
< \code{Lib} < \code{sliLib} < \code{Mod} < \code{sliCon} < \code{Con} <
\code{extCon}} \item{DoleLR}{Left-Right placement of Bob Dole (same scale as
selfLR): an ordered factor with levels \code{extLib} < \code{Lib} <
\code{sliLib} < \code{Mod} < \code{sliCon} < \code{Con} < \code{extCon}}
\item{PID}{Party identification: an ordered factor with levels strong
Democrat, \code{strDem} < weak Democrat, \code{weakDem} < independent
Democrat, \code{indDem} < independent independent\code{indind} < indepedent
Republican, \code{indRep} < waek Republican, \code{weakRep} < strong
Republican, \code{strRep}} \item{age}{Respondent's age in years}
\item{educ}{Respondent's education: an ordered factor with levels 8 years or
less, \code{MS} < high school dropout, \code{HSdrop} < high school diploma
or GED, \code{HS} < some College, \code{Coll} < Community or junior College
degree, \code{CCdeg} < BA degree, \code{BAdeg} < postgraduate degree,
\code{MAdeg}} \item{income}{Respondent's family income: an ordered factor
with levels \code{$3Kminus} < \code{$3K-$5K} < \code{$5K-$7K} <
\code{$7K-$9K} < \code{$9K-$10K} < \code{$10K-$11K} < \code{$11K-$12K} <
\code{$12K-$13K} < \code{$13K-$14K} < \code{$14K-$15K} < \code{$15K-$17K} <
\code{$17K-$20K} < \code{$20K-$22K} < \code{$22K-$25K} < \code{$25K-$30K} <
\code{$30K-$35K} < \code{$35K-$40K} < \code{$40K-$45K} < \code{$45K-$50K} <
\code{$50K-$60K} < \code{$60K-$75K} < \code{$75K-$90K} < \code{$90K-$105K} <
\code{$105Kplus}} \item{vote}{Expected vote in 1996 presidential election: a
factor with levels \code{Clinton} and \code{Dole}} }
}
\source{
Sapiro, Virginia, Steven J. Rosenstone, Donald R. Kinder, Warren E.
Miller, and the National Election Studies. AMERICAN NATIONAL ELECTION
STUDIES, 1992-1997: COMBINED FILE [Computer file]. 2nd ICPSR version. Ann
Arbor, MI: University of Michigan, Center for Political Studies [producer],
1999. Ann Arbor, MI: Inter-university Consortium for Political and Social
Research [distributor], 1999.
}
\description{
10 variable subset of the 1996 American National Election Study. Missing
values and "don't know" responses have been se deleted. Respondents
expressing a voting preference other than Clinton or Dole have been removed.
}
\references{
Found at \url{http://www.stat.washington.edu/}
}
\keyword{datasets}
