\name{odor}
\alias{odor}
\title{Odor of chemical by production settings}
\description{Data from an experiment to determine the effects of column
temperature, gas/liquid ratio and packing height in reducing
unpleasant odor of chemical product that was being sold for household
use}
\usage{data(odor)
}
\format{
  \describe{
    \item{\code{odor}}{ Odor score}
    \item{\code{temp}}{ Temperature coded as -1, 0 and 1}
    \item{\code{gas}}{ Gas/Liquid ratio coded as -1, 0 and 1}
    \item{\code{pack}}{ Packing height coded as -1, 0 and 1}
  }
}

\source{"Statistical Design and Analysis of Experiments" by P. John,
  Macmillan, 1971}
\keyword{datasets}
