\name{ozone}
\alias{ozone}
\title{Ozone readings in LA}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(ozone)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
