\name{parstum}
\alias{parstum}
\docType{data}
\title{Marijuana and parent alcohol and drug use}
\description{
  445 college
students were classified according to both frequency of marijuana use
and parental use of alcohol and psychoactive drugs.
}
\usage{data(parstum)}
\format{
  A data frame with 9 observations on the following 3 variables.
  \describe{
    \item{\code{parent}}{Number of parents using  drugs or alcohol - a factor with levels \code{Both} \code{Neither} \code{One}}
    \item{\code{student}}{Student usage of marijuana - a factor with levels \code{Never} \code{Occasional} \code{Regular}}
    \item{\code{count}}{the number of cases}
  }
}
\source{
  Ellis, Godfrey J. and Stone, Lorene H.  (1979)
  Marijuana Use in College: "An Evaluation of a Modeling Explanation"
  Youth and Society 10, 323-34
}
\keyword{datasets}
