\name{peanut}
\alias{peanut}
\title{
Carbon dioxide effects on peanut oil extraction
}
\description{
The \code{peanut} data frame has 16 rows and 6 columns.
Carbon dioxide effects on peanut oil extraction
}
\usage{
data(peanut)
}
\format{
This data frame contains the following columns:
\describe{
\item{\code{press}}{
CO2 pressure - two levels low=0, high=1
}
\item{\code{temp}}{
CO2 temperature - two levels low=0, high=1
}
\item{\code{moist}}{
peanut moisture - two levels low=0, high=1
}
\item{\code{flow}}{
CO2 flow rate - two levels low=0, high=1
}
\item{\code{size}}{
peanut particle size - two levels low=0, high=1
}
\item{\code{solubility}}{
the amount of oil that could dissolve in the CO2
}}}
\source{
  Kilgo, M (1989) "An Application of Fractional Factorial Experimental
  Designs"
  Quality Engineering, 1, 45-54
}
\keyword{datasets}

