\name{penicillin}
\alias{penicillin}
\title{Penicillin yields by block and treatment}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(penicillin)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
