\name{pima}
\alias{pima}
\title{Diabetes survey on Pima Indians}
\description{The National Institute of Diabetes and
Digestive and Kidney Diseases conducted a study on 768 adult female
Pima Indians living near Phoenix.
}
\usage{data(pima)
}
\format{The dataset contains the following variables
  \describe{
    \item{\code{pregnant}}{ Number of times pregnant}
    \item{\code{glucose}}{ Plasma glucose concentration at 2 hours in an
      oral glucose tolerance test}
    \item{\code{diastolic}}{ Diastolic blood pressure (mm Hg)}
    \item{\code{triceps}}{ Triceps skin fold thickness (mm)}
    \item{\code{insulin}}{ 2-Hour serum insulin (mu U/ml)}
    \item{\code{bmi}}{ Body mass index (weight in kg/(height in metres
      squared))}
    \item{\code{diabetes}}{ Diabetes pedigree function}
    \item{\code{age}}{ Age (years)}
    \item{\code{test}}{ test whether the patient shows signs of
      diabetes (coded 0 if negative, 1 if positive)}
  }
}
\source{ The data may
be obtained from
UCI Repository of machine learning databases at
\url{http://archive.ics.uci.edu/ml/}
}
\keyword{datasets}
