\name{pneumo}
\alias{pneumo}
\title{Miners suffering from pneumoconiosis}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(pneumo)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
