\name{potuse}
\alias{potuse}
\docType{data}
\title{Marijuana usage by youth}
\description{
  The National Youth Survey collected a sample of 11 to 17 year olds
  - 117 boys and 120 girls - asking questions about marijuana usage.
}
\usage{data(potuse)}
\format{
  A data frame with 486 observations on the following 7 variables.
  \describe{
    \item{\code{sex}}{1=Male, 2=Female}
    \item{\code{year.76}}{1=never used, 2=used no more than once a
      month, 3=used more than once a month in 1976}
    \item{\code{year.77}}{1=never used, 2=used no more than once a
      month, 3=used more than once a month in 1977}
    \item{\code{year.78}}{1=never used, 2=used no more than once a
      month, 3=used more than once a month in 1978}
    \item{\code{year.79}}{1=never used, 2=used no more than once a
      month, 3=used more than once a month in 1979}
    \item{\code{year.80}}{1=never used, 2=used no more than once a
      month, 3=used more than once a month in 1980}
    \item{\code{count}}{Number of cases in this category}
  }
}
\source{
  ICPSR, University of Michigan
}
\references{
  Lang J., McDonald, J and Smith P. (1999) "Association-Marginal
  Modeling of Mutlivariate Categorical Responses: A Maximum Likelihood
  Approach" JASA 94, 1161-
}
\keyword{datasets}
