\name{prostate}
\alias{prostate}
\title{Prostate cancer surgery}
\description{The \code{prostate} data frame has 97 rows and 9 columns.
A study on 97 men with prostate cancer who were due to receive
a radical prostatectomy.
}
\usage{
data(prostate)
}
\format{
This data frame contains the following columns:
\describe{

  \item{\code{lcavol}}{
    log(cancer volume)
  }
  \item{\code{lweight}}{
    log(prostate weight)
  }

  \item{\code{age}}{
      age
  }
  \item{\code{lbph}}{
      log(benign prostatic hyperplasia amount)
  }
  \item{\code{svi}}{
      seminal vesicle invasion
  }
  \item{\code{lcp}}{
      log(capsular penetration)
  }
  \item{\code{gleason}}{
      Gleason score
    }
  \item{\code{pgg45}}{
    percentage Gleason scores 4 or 5
  }
  \item{\code{lpsa}}{
    log(prostate specific antigen)
  }
  }}
\source{Andrews DF and Herzberg AM (1985): Data. New York: Springer-Verlag}
\keyword{datasets}
