\name{psid}
\alias{psid}
\docType{data}
\title{Panel Study of Income Dynamics subset}
\description{
  The Panel Study of Income Dynamics (PSID), begun in 1968, is a
  longitudinal study of a representative sample of U.S. individuals.
  The study is conducted at the Survey Research
Center, Institute for Social Research, University of Michigan and is
still continuing. The data represents a small subset of the total data.
}
\usage{data(psid)}
\format{
  A data frame with 1661 observations on the following 6 variables.
  \describe{
    \item{\code{age}}{age in 1968}
    \item{\code{educ}}{years of education}
    \item{\code{sex}}{sex of individual, \code{F} or \code{M}}
    \item{\code{income}}{annual income in dollars}
    \item{\code{year}}{calendar year}
    \item{\code{person}}{ID number for individual}
  }
}
\source{
Martha S. Hill, The Panel Study of Income Dynamics: A User's Guide, Sage Publications,
1992,Newbury Park, CA.
}
\keyword{datasets}
