\name{pulp}
\alias{pulp}
\title{
Brightness of paper pulp depending on shift operator
}
\description{
The \code{pulp} data frame has 20 rows and 2 columns.
Data comes from an experiment to test the paper brightness
depending on a shift operator.
}
\usage{
data(pulp)
}
\format{
This data frame contains the following columns:
\describe{
\item{\code{bright}}{
Brightness of the pulp as measured by a reflectance meter
}
\item{\code{operator}}{
  Shift operator a-d
}}}
\source{
  "Statistical techniques applied to production situations" F. Sheldon
  (1960)
  Industrial and Engineering Chemistry, 52, 507-509
}
\keyword{datasets}

