\name{pvc}
\alias{pvc}
\title{Production of PVC by operator and resin railcar}
\description{Data from an experiment to study factors affecting
  the production of the plastic PVC, 3 operators used 8 different devices called resin
railcars to produce PVC. For each of the 24 combinations, two samples
were produced.
}
\usage{data(pvc)
}
\format{Dataset contains the following variables
  \describe{
    \item{\code{psize}}{ Particle size}
    \item{\code{operator}}{ Operator number 1, 2 or 3}
    \item{\code{resin}}{ Resin railcar 1-8}
  }
}
\source{
R. Morris and E. Watson (1998) "A comparison of the techniques used to
evaluate the measurement process" Quality Engineering, 11, 213-219}
\keyword{datasets}
