\name{rabbit}
\alias{rabbit}
\title{Rabbit weight gain by diet and litter}
\description{A nutritionist studied the effects of six diets, 
on weight gain of domestic rabbits.  From past experience with sizes
of litters, it was felt that only 3 uniform rabbits could be selected
from each available litter. There were ten litters available forming
blocks of size three.
}
\usage{data(rabbit)
}
\format{The variables in the dataset were
  \describe{
    \item{\code{treat}}{ Diet a through f}

    \item{\code{gain}}{ Weight gain}

    \item{\code{block}}{ Block (10 litters)}
    }
}
\source{"Experimental Design and Analysis" by M. Lentner and T. Bishop,
  Valley Book Company, 1986
}
\keyword{datasets}
