\name{rats}
\alias{rats}
\title{Rat survival times by treatment and poison}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(rats)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
