\name{resceram}
\alias{resceram}
\title{
Shape and plate effects on current noise in resistors
}
\description{
The \code{resceram} data frame has 12 rows and 3 columns.
Shape and plate effects on current noise in resistors
}
\usage{
data(resceram)
}
\format{
This data frame contains the following columns:
\describe{
\item{\code{noise}}{
current noise
}
\item{\code{shape}}{
the geometrical shape of the resistor, A, B, C or D
}
\item{\code{plate}}{
  the ceramic plate on which the resistor was mounted. Only
  three resistors will fit on one plate.
}}}
\source{
  Natrella, M (1963) "Experimental Statistics" National Bureau
  of Standards Handbook 91, Gaithersburg MD.
}
\keyword{datasets}

