\name{seatpos}
\alias{seatpos}
\title{Car seat position depending driver size}
\description{Car drivers like to adjust the seat position for their own comfort.
Car designers would find it helpful to know where different drivers
will position the seat depending on their size and age. Researchers at
the HuMoSim laboratory at the University of Michigan collected data on
38 drivers. 
}
\usage{data(seatpos)}
\format{The dataset contains the following variables
  \describe{
    \item{\code{Age}}{ Age in years}
    \item{\code{Weight}}{ Weight in lbs}
    \item{\code{HtShoes}}{ Height in shoes in cm}
    \item{\code{Ht}}{ Height bare foot in cm}
    \item{\code{Seated}}{ Seated height in cm}
    \item{\code{Arm}}{ lower arm length in cm}
    \item{\code{Thigh}}{ Thigh length in cm}
    \item{\code{Leg}}{ Lower leg length in cm}
    \item{\code{hipcenter}}{ horizontal distance of the
      midpoint of the hips from a fixed location in the car in mm}
  }
}
\source{"Linear Models in R" by Julian Faraway, CRC Press, 2004
}
\keyword{datasets}
