\name{sexab}
\alias{sexab}
\title{Post traumatic stress disorder in abused adult females}
\description{The data for this example come from a study of the effects of
childhood sexual abuse on adult females. 45 women being treated at a clinic, who reported
childhood sexual abuse, were measured for post traumatic
stress disorder and childhood physical abuse both on
standardized scales. 31 women also being treated at the same clinic, who
did not report childhood sexual abuse were also measured. The full study
was more complex than reported here and so readers interested in the
subject matter should refer to the original article.

}
\usage{data(sexab)
}
\format{The variables in the dataset are
  \describe{

    \item{\code{cpa}}{ Childhood physical abuse on standard scale}

    \item{\code{ptsd}}{ Post-traumatic stress disorder on standard scale}
    \item{\code{csa}}{ Childhood sexual abuse - abused or not abused}
  }
}

\source{
  N. Rodriguez and S. Ryan and H. Vande Kemp and D. Foy (1997)
  "Postraumatic stress disorder in adult female survivors of
  childhood sexual abuse: A comparison study",
  Journal of Consulting and Clinical Pyschology, 65, 53-59
}
\keyword{datasets}
