% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{sexfun}
\alias{sexfun}
\title{Marital sex ratings}
\format{
A data frame with 16 observations on the following 3 variables.
\describe{ \item{y}{the count} \item{husband}{a factor with levels
\code{never} \code{fairly} \code{very} \code{always}} \item{wife}{a factor
with levels \code{never} \code{fairly} \code{very} \code{always}} }
}
\source{
Hout, M., Duncan, O. and Sobel M. (1987) Association and
heterogeneity: Structural models of similarities and differences.
Sociological Methods. 17, 145-184.
}
\description{
Data from a questionaire from 91 couples in the Tucson, Arizona area.
Subjects answered the question "Sex is fun for me and my partner". The
possible answers were "never or occasionally","fairly often","very often"
and "almost always"
}
\keyword{datasets}
