\name{snail}
\alias{snail}
\docType{data}
\title{
Snail production
}
\description{
A study was conducted to optimize snail
production for consumption. The percentage water content of the
tissues of snails grown under three different levels of relative humidity
 and two different temperatures was
recorded. For each combination, 4 snails were observed.
}
\usage{data(snail)}
\format{
  A data frame with 24 observations on the following 3 variables.
  \describe{
    \item{\code{water}}{percentage water content}
    \item{\code{temp}}{temperature in C}
    \item{\code{humid}}{relative humidity}
  }
}
\source{
Unknown
}
\examples{
data(snail)
## maybe str(snail) ; plot(snail) ...
}
\keyword{datasets}
