\name{solder}
\alias{solder}
\title{Solder skips in circuit board manufacture}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(solder)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
