% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{solv}
\alias{solv}
\title{Block design task testing child ability}
\format{
A data frame with 24 observations and 3 variables.
\describe{
\item{group}{Solution attempted first by row(r) or corner(c)}
\item{time}{Time taken to complete the task in seconds}
\item{eft}{Score on the embedded figures test which is a 
measure of difficulty in abstracting logical structure of 
a problem from its context.}
}
}
\source{
Statistical Modelling in GLIM (1989) 
M. Aitkin and D. Anderson and B. Francis and J. Hinde
Oxford University Press
}
\description{
Behavioural scientists at Macquarie University conducted an
experiment to test the time taken to perform a block design task with
24 fifth grade children (12 boys and 12 girls).
}
