\name{sono}
\alias{sono}
\title{
  Sonoluminescence
}
\description{
  The \code{sono} data frame has 16 rows and 8 columns.
  Sonoluminescence is the process of turning sound energy into light.
  An experiment was conducted to study factors affecting this process.
}
\usage{
data(sono)
}
\format{
This data frame contains the following columns:
\describe{
\item{\code{Intensity}}{
Sonoluminescent light intensity
}
\item{\code{Molarity}}{
  Amount of Solute. The coding is "low" for 0.10 mol and "high" for 0.33 mol.
}
\item{\code{Solute}}{
  Solute type. The coding is "low" for sugar and "high" for glycerol.
}
\item{\code{pH}}{
  The coding is "low" for 3 and "high" for 11.
}
\item{\code{Gas}}{
  Gas type in water. The coding is "low" for helium and "high" for air.
}
\item{\code{Water}}{
  Water depth. The coding is "low" for half and "high" for full.
}
\item{\code{Horn}}{
  Horn depth. The coding is "low" for 5 mm and "high" for 10 mm.
}
\item{\code{Flask}}{
  Flask clamping. The coding is "low" for unclamped and "high" for clamped.
}}}
\source{Eva Wilcox and Ken Inn of the NIST Physics Laboratory conducted
  this experiment during 1999 and published in NIST/SEMATECH e-Handbook of Statistical Methods, http://www.itl.nist.gov/div898/handbook/
}
\keyword{datasets}


