\name{spector}
\alias{spector}
\docType{data}
\title{Teaching methods in Economics}
\description{
  A study to determine the effectiveness of a new teaching method in Economics
}
\usage{data(spector)}
\format{
  A data frame with 32 observations on the following 4 variables.
  \describe{
    \item{grade}{1 = exam grades improved, 0 = not improved}
    \item{psi}{1 = student exposed to PSI (a new teach method), 0 = not exposed}
    \item{tuce}{a measure of ability when entering the class}
    \item{gpa}{grade point average}
  }
}
\source{
   Spector, L. and Mazzeo, M. (1980), "Probit Analysis and Economic
Education", Journal of Economic Education, 11, 37 - 44. 
}
\keyword{datasets}
