\name{speedo}
\alias{speedo}
\title{Speedometer cable shrinkage}
\description{Speedometer cables can be noisy because of
shrinkage in the plastic casing material. An experiment was conducted
to find out what caused shrinkage by screening a large number of
factors.  The engineers started with 15 different factors.

}
\usage{data(speedo)
}
\format{The dataset contains the following variables: (variables a-o
  are 2 level factors, coded "+" and "-" where "+" indicates a higher
  value where appropriate)
  \describe{
    \item{\code{a}}{ liner outer diameter}
    \item{\code{b}}{ liner die}
    \item{\code{c}}{ liner material}
    \item{\code{d}}{ liner line speed}
    \item{\code{e}}{ wire braid type}
    \item{\code{f}}{ braiding tension}
    \item{\code{g}}{ wire diameter}
    \item{\code{h}}{ liner tension}
    \item{\code{i}}{ liner temperature}
    \item{\code{j}}{ coating material}
    \item{\code{k}}{ coating die type}
    \item{\code{l}}{ melt temperature}
    \item{\code{m}}{ screen pack}
    \item{\code{n}}{ cooling method}
    \item{\code{o}}{ line speed}
    \item{\code{y}}{ percentage shrinkage per specimen}
  }
}
\source{G. P. Box and S. Bisgaard and C. Fung (1988) "An explanation and
  critque of Taguchi's contributions to quality engineering",
  Quality and reliability engineering international, 4, 123-131}
\keyword{datasets}

