\name{star}
\alias{star}
\title{Star light intensities and temperatures}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(star)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
