\name{stat500}
\alias{stat500}
\title{Scores for students in Stat500 class}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(stat500)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
