\name{suicide}
\alias{suicide}
\title{UK suicide data}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(suicide)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
