% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{turtle}
\alias{turtle}
\title{Incubation temperature and the sex of turtles}
\format{
A data frame with 15 observations on the following 3 variables.
\describe{ \item{temp}{temperature in degrees centigrade} \item{male}{number
of male turtles hatched} \item{female}{number of female turtles hatched} }
}
\source{
Beyond Traditional Statistical Methods Copyright 2000 D. Cook, P.
Dixon, W. M. Duckworth, M. S. Kaiser, K. Koehler, W. Q. Meeker and W. R.
Stephenson. Developed as part of NSF/ILI grant DUE9751644.
}
\description{
Incubation temperature can affect the sex of turtles. There are 3
independent replicates for each temperature.
}
\examples{

data(turtle)

}
\keyword{datasets}
