\name{tvdoctor}
\alias{tvdoctor}
\docType{data}
\title{
Life, TVs and Doctors
}
\description{
Life expectancy, doctors and televisions collected on 38 countries in 1993
}
\usage{data(tvdoctor)}
\format{
  A data frame with 38 observations on the following 3 variables.
  \describe{
    \item{\code{life}}{Life expectancy in years}
    \item{\code{tv}}{Number of people per television set}
    \item{\code{doctor}}{Number of people per doctor}
  }
}
\source{
Unknown, data for illustration purposes only
}
\examples{
data(tvdoctor)
## maybe str(tvdoctor) ; plot(tvdoctor) ...
}
\keyword{datasets}
