\name{twins}
\alias{twins}
\title{Twin IQs from Burt}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(twins)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
