\name{uncviet}
\alias{uncviet}
\title{UNC student opinion about the Vietnam war}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(uncviet)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
