% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{uswages}
\alias{uswages}
\title{Weekly wages of US male workers in 1988}
\format{
This data frame contains the following columns: \describe{
\item{wage}{ Real weekly wages in dollars (deflated by personal
consumption expenditures - 1992 base year)

} \item{educ}{ Years of education } \item{exper}{ Years of
experience } \item{race}{ 1 if Black, 0 if White (other races not in
sample) } \item{smsa}{ 1 if living in Standard Metropolitan
Statistical Area, 0 if not } \item{ne}{ 1 if living in the North
East } \item{mw}{ 1 if living in the Midwest } \item{we}{ 1
if living in the West } \item{so}{ 1 if living in the South }
\item{pt}{ 1 if working part time, 0 if not } }
}
\source{
Bierens, H.J., and D. Ginther (2001): "Integrated Conditional Moment
Testing of Quantile Regression Models", Empirical Economics 26, 307-324
}
\description{
The \code{uswages} data frame has 2000 rows and 10 columns. Weekly Wages for
US male workers sampled from the Current Population Survey in 1988.
}
\keyword{datasets}
