% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway.R
\name{VarCorrmat}
\alias{VarCorrmat}
\title{Helper function to convert VarCorr objects}
\usage{
VarCorrmat(varc, useScale, digits)
}
\arguments{
\item{varc}{a numeric vector}

\item{useScale}{a numeric vector}

\item{digits}{a numeric vector}
}
\value{
a matrix
}
\description{
Used internally when showing summary for lmer objects
}
\author{
copied from the arm package
}
\keyword{internal}
