\name{vision}
\alias{vision}
\title{Vision acuity tests}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(vision)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
