% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{weldstrength}
\alias{weldstrength}
\title{welding strength DOE}
\format{
A data frame with 16 observations on the following 10 variables.
\describe{ \item{Rods}{a 0-1 predictor} \item{Drying}{a 0-1 predictor}
\item{Material}{a 0-1 predictor} \item{Thickness}{a 0-1 predictor}
\item{Angle}{a 0-1 predictor} \item{Opening}{a 0-1 predictor}
\item{Current}{a 0-1 predictor} \item{Method}{a 0-1 predictor}
\item{Preheating}{a 0-1 predictor} \item{Strength}{The welding strength} }
}
\source{
G. Box and R. Meyer (1986) Dispersion effects from fractional
designs, Technometrics, 28, 19-27
}
\description{
An experiment to investigate factors affecting welding strength.
}
\keyword{datasets}
