% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{aatemp}
\alias{aatemp}
\title{Annual mean temperatures in Ann Arbor, Michigan}
\format{
A data frame with 115 observations on the following 2 variables.
\describe{ \item{year}{year from 1854 to 2000} \item{temp}{annual mean
temperatures in degrees F in Ann Arbor} }
}
\source{
United States Historical Climatology Network:
\url{http://www.ncdc.noaa.gov/oa/climate/research/ushcn/ushcn.html}
}
\description{
The data comes from the U.S. Historical Climatology Network.
}
\keyword{datasets}
