% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{airpass}
\alias{airpass}
\title{Airline passengers}
\format{
A data frame with 144 observations on the following 2 variables.
\describe{ \item{pass}{number of passengers in thousands}
\item{year}{the date as a decimal} }
}
\source{
Brown, R.G.(1962) Smoothing, Forecasting and Prediction of Discrete
Time Series.  Englewood Cliffs, N.J.: Prentice-Hall.
}
\description{
Monthly totals of airline passengers from 1949 to 1951
}
\details{
Well known time series example dataset
}
\examples{

data(airpass)
## maybe str(airpass) ; plot(airpass) ...

}
\references{
Box, G.E.P., Jenkins, G.M. and Reinsel, G.C. (1994) Time Series
Analysis, Forecasting and Control, 3rd edn.  Englewood Cliffs, N.J.:
Prentice-Hall.
}
\keyword{datasets}
