% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{amlxray}
\alias{amlxray}
\title{Match pair study for AML and Xray link}
\format{
A data frame with 238 observations on the following 11 variables.
\describe{ \item{ID}{a factor denoting the matched pairs}
\item{disease}{0=control, 1=case} \item{Sex}{ \code{F} or
\code{M}} \item{downs}{Presence of Downs syndrome: \code{no} or
\code{yes}} \item{age}{Age in years} \item{Mray}{Did the
mother ever have an Xray: \code{no} or \code{yes}} \item{MupRay}{Did
the mother have an Xray of the upper body during pregnancy: \code{no} or
\code{yes}} \item{MlowRay}{Did the mother have an Xray of the lower
body during pregnancy: \code{no} or \code{yes}} \item{Fray}{Did the
father ever have an Xray: \code{no} or \code{yes}} \item{Cray}{Did
the child ever have an Xray: \code{no} or \code{yes}}
\item{CnRay}{Total number of Xrays of the child \code{1}=none <
\code{2}=1 or 2 < \code{3}=3 or 4 < \code{4}= 5 or more} }
}
\source{
Chap T. Le (1998) "Applied Categorical Data Analysis" Wiley.
}
\description{
A matched case control study carried out to investigate the connection
between X-ray usage and acute myeloid leukemia in childhood. The pairs are
matched by age, race and county of residence.
}
\keyword{datasets}
