% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{anaesthetic}
\alias{anaesthetic}
\title{Time in minutes to eye opening after reversal of anaesthetic.}
\format{
A data frame with 80 observations on the following 2 variables.
\describe{ \item{breath}{time in minutes to start breathing
unassisted} \item{tgrp}{Four treatment groups \code{A} \code{B}
\code{C} \code{D}} }
}
\source{
Chatfield C. (1995) Problem Solving: A Statistician's Guide, 2ed
Chapman Hall.
}
\description{
A doctor at major London hospital compared the effects of 4 anaesthetics
used in major operations. 80 patients were divided into groups of 20.
}
\examples{

data(anaesthetic)
## maybe str(anaesthetic) ; plot(anaesthetic) ...

}
\keyword{datasets}
