\name{babyfood}
\alias{babyfood}
\title{Respiratory disease rates of babies fed in different ways}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(babyfood)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
