% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{babyfood}
\alias{babyfood}
\title{Respiratory disease rates of babies fed in different ways}
\format{
A data frame with 6 observations on the following 4 variables.
\describe{ \item{disease}{number with disease}
\item{nondisease}{number without disease} \item{sex}{a
factor with levels \code{Boy} \code{Girl}} \item{food}{a factor with
levels \code{Bottle} \code{Breast} \code{Suppl}} }
}
\source{
Payne, C. (1987). The GLIM System Release 3.77 Manual (2 ed.).
Oxford: Nu- merical Algorithms Group.
}
\description{
Study on infant respiratory disease, namely the proportions of children
developing bronchitis or pneumonia in their first year of life by type of
feeding and sex.
}
\examples{

data(babyfood)
## maybe str(babyfood) ; plot(babyfood) ...

}
\keyword{datasets}
