% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{beetle}
\alias{beetle}
\title{Beetles exposed to fumigant}
\format{
A data frame with 10 observations on the following 3 variables.
\describe{ \item{conc}{concentration of ethylene oxide in mg/l}
\item{affected}{number affected} \item{exposed}{number
exposed} }
}
\source{
Busvine (1938)
}
\description{
Grain beetles were exposed to ethylene oxide
}
\examples{

data(beetle)
## maybe str(beetle) ; plot(beetle) ...

}
\references{
Collet D. "Modelling Binary Data"
}
\keyword{datasets}
