\name{bliss}
\alias{bliss}
\title{Bliss insecticide data}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(bliss)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
