% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{bliss}
\alias{bliss}
\title{Insect mortality due to insecticide}
\format{
A data frame with 5 observations on the following 3 variables.
\describe{ \item{dead}{number dead} \item{alive}{number
alive} \item{conc}{concentration of insecticide} }
}
\source{
Bliss (1935). The calculation of the dosage-mortality curve. Annals
of Applied Biology 22, 134-167.
}
\description{
An experiment measuring death rates for insects, with 30 insects at each of
five treatment levels.
}
\examples{

data(bliss)
## maybe str(bliss) ; plot(bliss) ...

}
\keyword{datasets}
