% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{breaking}
\alias{breaking}
\title{Breaking strength of materials}
\format{
A data frame with 16 observations on the following 4 variables.
\describe{ \item{y}{The breaking strength of the component}
\item{operator}{the operator - a factor with levels \code{op1}
\code{op2} \code{op3} \code{op4}} \item{day}{the day of production -
a factor with levels \code{day1} \code{day2} \code{day3} \code{day4}}
\item{supplier}{the supplier of the raw material - a factor with
levels \code{A} \code{B} \code{C} \code{D}} }
}
\source{
Lentner M. and Bishop T. (1986) Experimental Design and Analysis,
Valley Book Company
}
\description{
An experiment was conducted to select the supplier of raw materials for
production of a component. The breaking strength of the component was the
objective of interest.  Four suppliers were considered. The four operators
can only produce one component each per day. A Latin square design was used.
}
\keyword{datasets}
