% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{butterfat}
\alias{butterfat}
\title{Butterfat content of milk by breed}
\format{
A data frame with 100 observations on the following 3 variables.
\describe{ \item{Butterfat}{butter fat content by percentage}
\item{Breed}{a factor with levels \code{Ayrshire} \code{Canadian}
\code{Guernsey} \code{Holstein-Fresian} \code{Jersey}} \item{Age}{a
factor with levels \code{2year} \code{Mature}} }
}
\source{
Sokal, R. R. and Rohlf, F. J. (1994) Biometry. W. H. Freeman, New York, third edition.
}
\description{
Average butterfat content (percentages) of milk for random samples of twenty
cows (ten two-year old and ten mature (greater than four years old)) from
each of five breeds. The data are from Canadian records of pure-bred dairy
cattle.
}
\examples{

data(butterfat)
## maybe str(butterfat) ; plot(butterfat) ...

}
\keyword{datasets}
