% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{cathedral}
\alias{cathedral}
\title{Cathedral nave heights and lengths in England}
\format{
A dataset with 25 cases \describe{
\item{style}{of the cathedral - romanesque or gothic}
\item{height}{in feet}
\item{width}{in feet}}
}
\source{
Weisberg, S. (2005). Applied Linear Regression, 3rd edition. New York: Wiley
}
\description{
Example Dataset from "Practical Regression and Anova"
}
\references{
Reference details may be found in "Practical Regression and
Anova" by Julian Faraway
}
\keyword{datasets}
