% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{cheddar}
\alias{cheddar}
\title{Taste of Cheddar cheese}
\format{
A data frame with 30 observations on the following 4 variables.
\describe{ \item{taste}{a subjective taste score}
\item{Acetic}{concentration of acetic acid (log scale)}
\item{H2S}{concentration of hydrogen sulfide (log scale)}
\item{Lactic}{concentration of lactic acid} }
}
\source{
David S. Moore and George P. McCabe (1993) Introduction to the Practice of Statistics,
W. H. Freeman and company, second edition.
}
\description{
In a study of cheddar cheese from the LaTrobe Valley of Victoria, Australia,
samples of cheese were analyzed for their chemical composition and were
subjected to taste tests. Overall taste scores were obtained by combining
the scores from several tasters.
}
\examples{

data(cheddar)
## maybe str(cheddar) ; plot(cheddar) ...

}
\keyword{datasets}
