% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{chicago}
\alias{chicago}
\title{Chicago insurance redlining}
\format{
This dataframe contains the following columns \describe{

\item{race}{ racial composition in percent minority }
\item{fire}{ fires per 100 housing units } \item{theft}{
theft per 1000 population } \item{age}{ percent of housing units
built before 1939 } \item{involact}{ new FAIR plan policies and
renewals per 100 housing units } \item{income}{ median family income
in thousands of dollars} \item{side}{ North or South side of
Chicago}

}
}
\source{
Adapted from "Data : A Collection of Problems from Many Fields for
the Student and Research Worker" by D. Andrews and A. Herzberg published by
Springer-Verlag, in 1985
}
\description{
Data from a 1970's study on the relationship between insurance redlining in
Chicago and racial composition, fire and theft rates, age of housing and
income in 47 zip codes.
}
\keyword{datasets}
