% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{chiczip}
\alias{chiczip}
\title{Chicago zip codes north-south}
\format{
\describe{
\item{chiczip}{takes the values "n" (north) and "s" south}
}
}
\description{
Complements the chicago and chmiss datasets by dividing the zip codes into
north and south
}
\references{
Reference details may be found in "Practical Regression and
Anova" by Julian Faraway
}
\seealso{
chicago
}
\keyword{datasets}
