\name{clot}
\alias{clot}
\title{Blood clotting times}
\description{The clotting times of blood for plasma diluted with nine
  different percentage concentrations with prothrombin-free plasma
}
\usage{data(clot)
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{\code{time}}{time in seconds to clot}
    \item{\code{conc}}{concentration in percent}
    \item{\code{lot}}{lot number - either one or two}
  }
  }
\source{
Hurn et al (1945)
}
\references{
 Nelder & McCullagh (1989) Generalized Linear Models (2ed)
 }
\keyword{datasets}
