% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{clot}
\alias{clot}
\title{Blood clotting times}
\format{
This data frame contains the following columns: \describe{
\item{time}{time in seconds to clot}
\item{conc}{concentration in percent} \item{lot}{lot number
- either one or two} }
}
\source{
Hurn et al (1945)
}
\description{
The clotting times of blood for plasma diluted with nine different
percentage concentrations with prothrombin-free plasma
}
\references{
Nelder & McCullagh (1989) Generalized Linear Models (2ed)
}
\keyword{datasets}
