% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{composite}
\alias{composite}
\title{Strength of a thermoplastic composite depending on two factors}
\format{
This data frame contains the following columns: \describe{
\item{strength}{ interply bond strength of the composite }
\item{laser}{ laser power at 40, 50 or 60W } \item{tape}{
tape speed, slow=6.42 m/s, medium=13m/s and fast=27m/s }}
}
\source{
Mazumdar, S and Hoa S (1995) "Application of a Taguchi Method for
Process enhancement of an online consolidation technique" Composites 26,
669-673
}
\description{
The \code{composite} data frame has 9 rows and 3 columns. Data comes from an
experiment to test the strength of a thermoplastic composite depending on
the power of a laser and speed of a tape.
}
\keyword{datasets}
