% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{cornnit}
\alias{cornnit}
\title{Corn yields from nitrogen application}
\format{
A data frame with 44 observations on the following 2 variables.
\describe{ \item{yield}{corn yield in bushels per acre}
\item{nitrogen}{pounds per acre} }
}
\source{
Unknown
}
\description{
The relationship between corn yield (bushels per acre) and nitrogen (pounds
per acre) fertilizer application were studied in Wisconsin.
}
\keyword{datasets}
