% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{cpd}
\alias{cpd}
\title{Projected and actual sales of 20 consumer products}
\format{
A data frame with 20 observations on the following 2 variables.
\describe{ \item{projected}{projected sales in dollars} \item{actual}{actual
sales in dollars} }
}
\source{
G. Whitmore (1986) "Inverse Gaussian Ratio Estimation" Applied
Statistics, 35, 8-15.
}
\description{
Projected and actual sales of 20 consumer products. Data have been disguised
from original form.
}
\keyword{datasets}
