% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway.R
\name{Cpplot}
\alias{Cpplot}
\title{Cp plot}
\usage{
Cpplot(cp)
}
\arguments{
\item{cp}{A leaps object returned from leaps()}
}
\value{
none
}
\description{
Makes a Cp plot
}
\details{
Requires leaps package
}
\seealso{
leaps()
}
\author{
Julian Faraway
}
\keyword{regression}
