\name{ctsib}
\alias{ctsib}
\title{Effects of surface and vision on balance}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(ctsib)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
