% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{death}
\alias{death}
\title{Death penalty in Florida 1977}
\format{
A data frame with 8 observations on the following 4 variables.
\describe{ \item{y}{a numeric vector} \item{penalty}{Did the
subject recieve the death penalty?  \code{no} or \code{yes}}
\item{victim}{Was the victim \code{b}lack or \code{w}hite?}
\item{defend}{Was the defendent \code{b}lack or \code{w}hite?} }
}
\source{
Radelet M. (1981) Racial characteristics and the imposition of the
death penalty. Amer. Sociol. Rev. \bold{46} 918-927.
}
\description{
Data on 326 defendents in homicide indictments in 20 Florida counties during
1976-77.
}
\references{
Agresti A. (1990) Categorical Data Analysis, Wiley.
}
\keyword{datasets}
