% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{denim}
\alias{denim}
\title{Denim wastage by supplier}
\format{
A data frame with 95 observations on the following 2 variables.
\describe{ \item{waste}{percentage wastage}
\item{supplier}{a factor with levels \code{1} \code{2} \code{3}
\code{4} \code{5}} }
}
\source{
Unknown
}
\description{
Five suppliers cut denim material for a jeans manufacturer. An algorithm is
used to estimate how much material will be wasted given the dimensions of
the material supplied. Typically, a supplier wastes more material than the
target based on the algorithm although occasionally they waste less. The
percentage of waste relative to target was collected weekly for the 5
suppliers. In all, 95 observations were recorded.
}
\examples{

data(denim)
## maybe str(denim) ; plot(denim) ...

}
\keyword{datasets}
