% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes and obesity, cardiovascular risk factors}
\format{
A data frame with 403 observations on the following 19 variables.
\describe{ \item{id}{Subject ID} \item{chol}{Total
Cholesterol} \item{stab.glu}{Stabilized Glucose}
\item{hdl}{High Density Lipoprotein}
\item{ratio}{Cholesterol/HDL Ratio}
\item{glyhb}{Glycosolated Hemoglobin} \item{location}{County
- a factor with levels \code{Buckingham} \code{Louisa}}
\item{age}{age in years} \item{gender}{a factor with levels
\code{male} \code{female}} \item{height}{height in inches}
\item{weight}{weight in pounds} \item{frame}{a factor with
levels \code{small} \code{medium} \code{large}} \item{bp.1s}{First
Systolic Blood Pressure} \item{bp.1d}{First Diastolic Blood
Pressure} \item{bp.2s}{Second Systolic Blood Pressure}
\item{bp.2d}{Second Diastolic Blood Pressure}
\item{waist}{waist in inches} \item{hip}{hip in inches}
\item{time.ppn}{Postprandial Time (in minutes) when Labs were Drawn}
}
}
\source{
Willems JP, Saunders JT, DE Hunt, JB Schorling: Prevalence of
coronary heart disease risk factors among rural blacks: A community-based
study. Southern Medical Journal 90:814-820; 1997
}
\description{
403 African Americans were interviewed in a study to understand the
prevalence of obesity, diabetes, and other cardiovascular risk factors in
central Virginia.
}
\details{
Glycosolated hemoglobin greater than 7.0 is usually taken as a positive
diagnosis of diabetes
}
\references{
Schorling JB, Roach J, Siegel M, Baturka N, Hunt DE, Guterbock
TM, Stewart HL: A trial of church-based smoking cessation interventions for
rural African Americans. Preventive Medicine 26:92-101; 1997
}
\keyword{datasets}
