\name{drugpsy}
\alias{drugpsy}
\docType{data}
\title{Choice of drug treatment for psychiatry patients}
\description{
  A sample of psychiatry patients were cross-classified by their
  diagnosis and whether a drug treatment was prescribed.
}
\usage{data(drugpsy)}
\format{
  A data frame with 10 observations on the following 3 variables.
  \describe{
    \item{\code{y}}{the number of patients}
    \item{\code{diagnosis}}{a factor with levels \code{Affective.Disorder} \code{Neurosis} \code{Personality.Disorder} \code{Schizophrenia} \code{Special.Symptoms}}
    \item{\code{drug}}{a factor with levels \code{no} \code{yes}}
  }
}
}
\source{
  Helmes E. and Fekken G. (1986) Effects of psychotropic drugs and
  psychiatric illness on vocational aptitude and interest
  assessment. J. Clin. Psychol. \bold{42} 569-576
}
\references{
  Agresti A. (1990) "Categorical Data Analysis" Wiley
}
\keyword{datasets}
