% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{drugpsy}
\alias{drugpsy}
\title{Choice of drug treatment for psychiatry patients}
\format{
A data frame with 10 observations on the following 3 variables.
\describe{ \item{y}{the number of patients}
\item{diagnosis}{a factor with levels \code{Affective.Disorder}
\code{Neurosis} \code{Personality.Disorder} \code{Schizophrenia}
\code{Special.Symptoms}} \item{drug}{a factor with levels \code{no}
\code{yes}} }
}
\source{
Helmes E. and Fekken G. (1986) Effects of psychotropic drugs and
psychiatric illness on vocational aptitude and interest assessment. J. Clin.
Psychol. \bold{42} 569-576
}
\description{
A sample of psychiatry patients were cross-classified by their diagnosis and
whether a drug treatment was prescribed.
}
\references{
Agresti A. (1990) "Categorical Data Analysis" Wiley
}
\keyword{datasets}
