% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{eco}
\alias{eco}
\title{Ecological regression example}
\format{
This dataframe contains the following columns \describe{
\item{usborn}{ Percentage of population born in the United States}
\item{income}{ Per capita annual income in dollars}
\item{home}{ Percentage born in state} \item{pop}{
Population of state } }
}
\source{
US Bureau of the Census
}
\description{
Relationship between 1998 per capita income dollars from all sources and the
proportion of legal state residents born in the United States in 1990 for
each of the 50 states plus the District of Columbia
}
\keyword{datasets}
