% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{eggprod}
\alias{eggprod}
\title{Treatment and block effects on egg production}
\format{
This data frame contains the following columns: \describe{
\item{treat}{ Three treatments: O, E or F } \item{block}{
Four blocks labeled 1-4 } \item{eggs}{ Number of eggs produced }}
}
\source{
Mead, R., R.N. Curnow, and A.M. Hasted. 1993. Statistical Methods in
Agriculture and Experimental Biology. Chapman and Hall, London, p. 64. 1993
}
\description{
The \code{eggprod} data frame has 12 rows and 3 columns.  Six pullets were
placed into each of 12 pens. Four blocks were formed from groups of 3 pens
based on location. Three treatments were applied. The number of eggs
produced was recorded
}
\keyword{datasets}
