\name{epilepsy}
\alias{epilepsy}
\title{Seizure rates of epileptics under treatment}
\description{Example Dataset from "Practical Regression and Anova"
}
\usage{data(epilepsy)
}
\format{See for yourself
}
\source{See Reference
}
\references{Reference details may be found in "Practical Regression and
  Anova" by Julian Faraway
  }
\keyword{datasets}
