% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{esdcomp}
\alias{esdcomp}
\title{Complaints about emergency room doctors}
\format{
A data frame with 44 observations on the following 6 variables.
\describe{ \item{visits}{the number of patient visits}
\item{complaints}{the number of complaints}
\item{residency}{is the doctor in residency training \code{N} or
\code{Y}} \item{gender}{gender of doctor \code{F} or \code{M}}
\item{revenue}{dollars per hour earned by the doctor}
\item{hours}{total number of hours worked} }
}
\source{
Chap T. Le (1998) "Applied Categorical Data Analysis" Wiley
}
\description{
Data was recorded on 44 doctors working in an emergency service at a
hospital to study the factors affecting the number of complaints received.
}
\keyword{datasets}
