% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{exa}
\alias{exa}
\title{Simulated non-parametric regression data}
\format{
A data frame with 256 observations on the following 3 variables.
\describe{ \item{x}{input} \item{y}{response}
\item{m}{true value} }
}
\source{
Haerdle, W. (1991). Smoothing Techniques with Implementation in S.
New York:Springer.
}
\description{
True function is f(x)=sin^3(2pi x^3).
}
\examples{

data(exa)
## maybe str(exa) ; plot(exa) ...

}
\keyword{datasets}
