% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{eyegrade}
\alias{eyegrade}
\title{grading of eye pairs for distance vision}
\format{
A data frame with 16 observations on the following 3 variables.
\describe{ \item{y}{the observed count} \item{right}{rated vision in the
right eye - a factor with levels \code{best} \code{second} \code{third}
\code{worst}} \item{left}{rated vision in the left eye - a factor with
levels \code{best} \code{second} \code{third} \code{worst}} }
}
\source{
A. Stuart (1955) A test for homogeneity of the marginal
distributions in a two-way classification, Biometrika, 42, 412-416.
}
\description{
A sample of women are rated for the performance of distance vision in each
eye.
}
\keyword{datasets}
