% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{fat}
\alias{fat}
\title{Percentage of Body Fat and Body Measurements}
\format{
A data frame with 252 observations on the following 18 variables.
\describe{ \item{brozek}{ Percent body fat using Brozek's equation,
457/Density - 414.2} \item{siri}{ Percent body fat using Siri's
equation, 495/Density - 450} \item{density}{Density (gm/$cm^3$)}
\item{age}{ Age (yrs)} \item{weight}{ Weight (lbs)}
\item{height}{ Height (inches)} \item{adipos}{ Adiposity
index = Weight/Height$^2$ (kg/$m^2$)} \item{free}{ Fat Free Weight =
(1 - fraction of body fat) * Weight, using Brozek's formula (lbs)}
\item{neck}{ Neck circumference (cm)} \item{chest}{ Chest
circumference (cm)} \item{abdom}{ Abdomen circumference (cm) at the
umbilicus and level with the iliac crest} \item{hip}{ Hip
circumference (cm)} \item{thigh}{ Thigh circumference (cm)}
\item{knee}{ Knee circumference (cm)} \item{ankle}{ Ankle
circumference (cm)} \item{biceps}{ Extended biceps circumference
(cm)} \item{forearm}{Forearm circumference (cm)}
\item{wrist}{ Wrist circumference (cm) distal to the styloid
processes} }
}
\source{
Johnson R. Journal of Statistics Education v.4, n.1 (1996)
}
\description{
Age, weight, height, and 10 body circumference measurements are recorded for
252 men. Each man's percentage of body fat was accurately estimated by an
underwater weighing technique.
}
\keyword{datasets}
