% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{fortune}
\alias{fortune}
\title{Billionaires' wealth and age}
\format{
A data frame with 232 observations on the following 3 variables.
\describe{
\item{wealth}{Billions of dollars}
\item{age}{age in years} \item{region}{a factor with levels \code{A}, Asia,
\code{E}, Europe, \code{M}, Middle East, \code{O} Other, \code{U} USA} }
}
\source{
Fortune magazine
}
\description{
Fortune magazine publishes a f the world's billionaires each year. The
1992 list includes 233 individuals. Their wealth, age, and geographic
location (Asia, Europe, Middle East, United States, and Other) are reported.
}
\examples{

data(fortune)
## maybe str(fortune) ; plot(fortune) ...

}
\keyword{datasets}
