% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{fpe}
\alias{fpe}
\title{1981 French Presidential Election}
\format{
This dataframe contains the following columns (vote totals are in
thousands) \describe{ \item{list("EI}{ Electeur Inscrits (registered
voters)}

\item{A}{ Voters for Mitterand in the first round} \item{B}{
Voters for Giscard in the first round} \item{C}{ Voters for Chirac
in the first round} \item{D}{ Voters for Communists in the first
round} \item{E}{ Voters for Ecology party in the first round}
\item{F}{ Voters for party F in the first round} \item{G}{
Voters for party G in the first round} \item{H}{ Voters for party H
in the first round} \item{I}{ Voters for party I in the first round}
\item{J}{ Voters for party J in the first round} \item{K}{
Voters for party K in the first round} \item{A2}{ Voters for
Mitterand in the second round} \item{B2}{ Voters for party Giscard
in the second round} \item{N}{ Difference between the number of
voters in the second round and in the first round}

}
}
\source{
"The Teaching of Practical Statistics" by C.W. Anderson and R.M.
Loynes, Wiley,1987
}
\description{
Elections for the French presidency proceed in two rounds. In 1981, there
were 10 candidates in the first round. The top two candidates then went on
to the second round, which was won by Francois Mitterand over Valery
Giscard-d'Estaing. The losers in the first round can gain political favors
by urging their supporters to vote for one of the two fina Since
voting is private, we cannot know how these votes were transferred, we might
hope to infer from the published vote totals how this might have happened.
Data is given for vote totals in every fourth department of France:
}
\keyword{datasets}
