% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway.R
\name{fround}
\alias{fround}
\alias{pfround}
\title{Formating the Rounding of Numbers}
\usage{
fround(x, digits)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{integer indicating the precision to be used.}
}
\description{
\code{fround} rounds the values in its first argument to the specified
number of decimal places with surrounding quotes.
}
\details{
\code{pfround} rounds the values in its first argument to the specified
number of decimal places without surrounding quotes.
}
\examples{

    x <- 3.1415926
    fround(x, digits=2)
    pfround(x, digits=2)

}
\references{
Copied from the \code{arm} package
}
\seealso{
\code{\link{round}}
}
\author{
Andrew Gelman; Yu-Sung Su
}
\keyword{manip}
\keyword{print}
