% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{gala}
\alias{gala}
\title{Species diversity on the Galapagos Islands}
\format{
The dataset contains the following variables \describe{
\item{Species}{ the number of plant species found on the island}
\item{Endemics}{ the number of endemic species} \item{Area}{
the area of the island (km$^2$)} \item{Elevation}{ the highest
elevation of the island (m)} \item{Nearest}{ the distance from the
nearest island (km)} \item{Scruz}{ the distance from Santa Cruz
island (km)} \item{Adjacent}{ the area of the adjacent island
(square km)} }
}
\source{
M. P. Johnson and P. H. Raven (1973) "Species number and endemism:
The Galapagos Archipelago revisited" Science, 179, 893-895
}
\description{
There are 30 Galapagos islands and 7 variables in the dataset. The
relationship between the number of plant species and several geographic
variables is of interest. The original dataset contained several missing
values which have been filled for convenience. See the \code{galamiss}
dataset for the original version.
}
\keyword{datasets}
