% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{gammaray}
\alias{gammaray}
\title{Xray decay from a gamma ray burst}
\format{
A data frame with 63 observations on the following 3 variables.
\describe{ \item{time}{in seconds since burst}
\item{flux}{X-ray flux in units of 10^-11 erg/cm2/s, 2-10 keV}
\item{error}{measurement error of the flux based on detector
signal-to-noise values} }
}
\source{
A. J. Blustin and 64 coauthors, Astrophys. J. 637, 901-913 2006.
Available at http://arxiv.org/abs/astro-ph/0507515.
}
\description{
The X-ray decay light curve of Gamma ray burst 050525a obtained with the
X-Ray Telescope (XRT) on board the Swift satellite. The dataset has 63
brightness measurements in the 0.4-4.5 keV spectral band at times ranging
from 2 minutes to 5 days after the burst.
}
\examples{

data(gammaray)
## maybe str(gammaray) ; plot(gammaray) ...

}
\keyword{datasets}
