% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{gavote}
\alias{gavote}
\title{Undercounted votes in Georgia in 2000 presidential election}
\format{
A data frame with 159 observations on the following 10 variables.
Each case represents a county in Georgia.  \describe{
\item{equip}{The voting equipment used: \code{LEVER}, \code{OS-CC}
(optical, central count), \code{OS-PC} (optical, precinct count)
\code{PAPER}, \code{PUNCH}} \item{econ}{economic status of county:
\code{middle} \code{poor} \code{rich}} \item{perAA}{percent of
African Americans in county} \item{rural}{indicator of whether
county is \code{rural} or \code{urban}} \item{atlanta}{indicator of
whether county is in \code{Atlanta} or not: \code{notAtlanta}}
\item{gore}{number of votes for Gore} \item{bush}{number of
votes for Bush} \item{other}{number of votes for other candidates}
\item{votes}{number of votes} \item{ballots}{number of
ballots} }
}
\source{
Meyer M. (2002) Uncounted Votes: Does Voting Equipment Matter?
Chance, 15(4), 33-38
}
\description{
The data comes from the US presidential election in the state of Georgia.
The undercount is the difference between the number of ballots cast and
votes recorded. Voters may have chosen not to vote for president, voted for
more than one candidate (disqualified) or the equipment may have failed to
register their choice.
}
\keyword{datasets}
