% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway.R
\name{halfnorm}
\alias{halfnorm}
\title{Half Normal Plot}
\usage{
halfnorm(
  x,
  nlab = 2,
  labs = as.character(1:length(x)),
  ylab = "Sorted Data",
  ...
)
}
\arguments{
\item{x}{a numeric vector}

\item{nlab}{number of points to label}

\item{labs}{labels for points}

\item{ylab}{label for Y-axis}

\item{...}{arguments passed to plot()}
}
\value{
none
}
\description{
Makes a half-normal plot
}
\examples{

halfnorm(runif(10))

}
\seealso{
qqnorm
}
\author{
Julian Faraway
}
