% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{hemoglobin}
\alias{hemoglobin}
\title{Treatment of insulin dependent diabetic children}
\format{
A data frame with 16 observations on the following 5 variables.
\describe{ \item{age}{age in years} \item{sex}{a factor with
levels \code{F} \code{M}} \item{treatment}{a factor with levels
\code{N} \code{S}} \item{pre}{measured value of hemoglobin before
treatment} \item{post}{measured value of hemoglobin after treatment}
}
}
\source{
Unknown
}
\description{
16 insulin-dependent diabetic children were enrolled in a study involving a
new treatment. 8 children received the new treatment(N) while the other 8
received the standard treatment(S).  The age and sex of the child was
recorded along with the measured value of gycosolated hemoglobin both before
and after treatment.
}
\examples{

data(hemoglobin)
## maybe str(hemoglobin) ; plot(hemoglobin) ...

}
\keyword{datasets}
