% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{hips}
\alias{hips}
\title{Ankylosing Spondylitis}
\format{
A data frame with 78 observations on the following 7 variables.
\describe{ \item{fbef}{flexion angle before}
\item{faft}{flexion angle after} \item{rbef}{rotation angle
before} \item{raft}{rotation angle after}
\item{grp}{treatment group - a factor with levels \code{control}
\code{treat}} \item{side}{side of the body - a factor with levels
\code{right} \code{left}} \item{person}{id for the individual} }
}
\source{
Chatfield C. (1995) Problem Solving: A Statistician's Guide, 2ed
Chapman Hall.
}
\description{
Data from Royal Mineral Hospital in Bath. AS is a chronic form of arthritis.
A study conducted to determine whether daily stretching of the hip tissues
would improve mobility.  39 ``typical'' AS patients were randomly allocated
to control (standard treatment) group or the treatment group in a 1:2 ratio.
Responses were flexion and rotation angles at the hip measured in degrees.
Larger numbers indicate more flexibility.
}
\examples{

data(hips)
## maybe str(hips) ; plot(hips) ...

}
\keyword{datasets}
