\name{hormone}
\alias{hormone}
\docType{data}
\title{
Hormone concentrations in gay and straight men
}
\description{
Urinary androsterone
(androgen) and etiocholanolone (estrogen) values were recorded from 26
healthy males.
}
\usage{data(hormone)}
\format{
  A data frame with 26 observations on the following 3 variables.
  \describe{
    \item{\code{androgen}}{concentration}
    \item{\code{estrogen}}{concentration}
    \item{\code{orientation}}{sexual orientation with levels \code{g} \code{s}}
  }
}
\source{
Margolese, M. (1970). Homosexuality: A new endocrine correlate. Hormones
and Behavior 1, 151-155.
}
\references{
Hand, D. (1981). Discrimination and Classification. Chichester, UK: Wiley.
}
\examples{
data(hormone)
## maybe str(hormone) ; plot(hormone) ...
}
\keyword{datasets}
