% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{hormone}
\alias{hormone}
\title{Hormone concentrations in gay and straight men}
\format{
A data frame with 26 observations on the following 3 variables.
\describe{ \item{androgen}{concentration}
\item{estrogen}{concentration} \item{orientation}{sexual
orientation with levels \code{g} \code{s}} }
}
\source{
Margolese, M. (1970). Homosexuality: A new endocrine correlate.
Hormones and Behavior 1, 151-155.
}
\description{
Urinary androsterone (androgen) and etiocholanolone (estrogen) values were
recorded from 26 healthy males.
}
\examples{

data(hormone)
## maybe str(hormone) ; plot(hormone) ...

}
\references{
Hand, D. (1981). Discrimination and Classification. Chichester,
UK: Wiley.
}
\keyword{datasets}
