% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{hprice}
\alias{hprice}
\title{Housing prices in US cities 86-94}
\format{
A data frame with 324 observations on the following 8 variables.
\describe{ \item{narsp}{natural log average sale price in thousands
of dollars} \item{ypc}{average per capita income}
\item{perypc}{percentage growth in per capita income}
\item{regtest}{Regulatory environment index (high values = more
regulations)} \item{rcdum}{Rent control - a factor with levels
\code{0}=no \code{1}=yes} \item{ajwtr}{Adjacent to a coastline - a
factor with levels \code{0}=no \code{1}=yes} \item{msa}{indicator
for the MSA} \item{time}{Year 1=1986 to 9=1994} }
}
\source{
Longitudinal and Panel Data: Analysis and Applications in the Social
Sciences, by Edward W. Frees, Cambridge University Press, August 2004.
}
\description{
Data on housing prices in 36 US metropolitan statistical areas (MSAs) over 9
years from 1986-1994 were collected.
}
\keyword{datasets}
