% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{hsb}
\alias{hsb}
\title{Career choice of high school students}
\format{
A data frame with 200 observations on the following 11 variables.
\describe{ \item{id}{ID of student} \item{gender}{a factor
with levels \code{female} \code{male}} \item{race}{a factor with
levels \code{african-amer} \code{asian} \code{hispanic} \code{white}}
\item{ses}{socioeconomic class - a factor with levels \code{high}
\code{low} \code{middle}} \item{schtyp}{school type - a factor with
levels \code{private} \code{public}} \item{prog}{choice of high
school program - a factor with levels \code{academic} \code{general}
\code{vocation}} \item{read}{reading score}
\item{write}{writing score} \item{math}{math score}
\item{science}{science score} \item{socst}{social science
score} }
}
\source{
National Education Longitudinal Studies (NELS) program of the
National Center for Education Statistics (NCES).
}
\description{
Data was collected as a subset of the "High School and Beyond" study
conducted by the National Education Longitudinal Studies (NELS) program of
the National Center for Education Statistics (NCES).
}
\details{
One purpose of the study was to determine which factors are related to the
choice of the type of program, academic, vocational or general, that the
students pursue in high school.
}
\keyword{datasets}
