% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway.R
\name{ilogit}
\alias{ilogit}
\title{Inverse Logit Transformation}
\usage{
ilogit(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
exp(x)/(1+exp(x))
}
\description{
Computes the inverse logit transformation
}
\examples{

ilogit(1:3)
#[1] 0.7310586 0.8807971 0.9525741

}
\seealso{
logit
}
\author{
Julian Faraway
}
