% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{infmort}
\alias{infmort}
\title{Infant mortality according to income and region}
\format{
This data frame contains the following columns: \describe{

\item{region}{ Region of the world, Africa, Europe, Asia or the
Americas } \item{income}{ Per capita annual income in dollars }

\item{mortality}{ Infant mortality in deaths per 1000 births }
\item{oil}{ Does the country export oil or not?  } }
}
\source{
Unknown
}
\description{
The \code{infmort} data frame has 105 rows and 4 columns.  The infant
mortality in regions of the world may be related to per capita income and
whether oil is exported. The dataset is not recent.
}
\keyword{datasets}
