% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{insulgas}
\alias{insulgas}
\title{Effects of insulation on gas consumption}
\format{
A data frame with 44 observations on the following 3 variables.
\describe{ \item{Insulate}{a factor with levels \code{After}
\code{Before}} \item{Temp}{Outside temperature}
\item{Gas}{Weekly consumption in 1000 cubic feet} }
}
\source{
MASS package as whiteside
}
\description{
Data on natural gas usage in a house.  The weekly gas consumption (in 1000
cubic feet) and the average outside temperature (in degrees Celsius) was
recorded for 26 weeks before and 30 weeks after cavity-wall insulation had
been installed.  The house thermostat was set at 20C throughout.
}
\examples{

data(insulgas)
## maybe str(insulgas) ; plot(insulgas) ...

}
\keyword{datasets}
