% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{jsp}
\alias{jsp}
\title{Junior School Project}
\format{
A data frame with 3236 observations on the following 9 variables.
\describe{ \item{school}{50 schools code 1-50}
\item{class}{a factor with levels \code{1} \code{2} \code{3}
\code{4}} \item{gender}{a factor with levels \code{boy} \code{girl}}
\item{social}{class of the father I=1; II=2; III nonmanual=3; III
manual=4; IV=5; V=6; Long-term unemployed=7; Not currently employed=8;
Father absent=9} \item{raven}{test score} \item{id}{student
id coded 1-1402} \item{english}{score on English}
\item{math}{score on Maths} \item{year}{year of school} }
}
\source{
Mortimore, P., P. Sammons, L. Stoll, D. Lewis, and R. Ecob (1988).
School Matters. Wells, UK: Open Books.
}
\description{
Junior School Project collected from primary (U.S. term is elementary)
schools in inner London.
}
\examples{

data(jsp)
## maybe str(jsp) ; plot(jsp) ...

}
\references{
Goldstein, H. (1995). Multilevel Statistical Models (2 ed.).
London: Arnold.
}
\keyword{datasets}
