% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{kanga}
\alias{kanga}
\title{Kangaroo skull measurements}
\format{
A data frame with 148 observations on the following 20 variables.
\describe{ \item{species}{a factor with levels \code{fuliginosus}
\code{giganteus} \code{melanops}} \item{sex}{a factor with levels
\code{Female} \code{Male}} \item{basilar.length}{a numeric vector}
\item{occipitonasal.length}{a numeric vector}
\item{palate.length}{a numeric vector} \item{palate.width}{a
numeric vector} \item{nasal.length}{a numeric vector}
\item{nasal.width}{a numeric vector}
\item{squamosal.depth}{a numeric vector}
\item{lacrymal.width}{a numeric vector}
\item{zygomatic.width}{a numeric vector}
\item{orbital.width}{a numeric vector}
\item{rostral.width}{a numeric vector}
\item{occipital.depth}{a numeric vector}
\item{crest.width}{a numeric vector}
\item{foramina.length}{a numeric vector}
\item{mandible.length}{a numeric vector}
\item{mandible.width}{a numeric vector}
\item{mandible.depth}{a numeric vector}
\item{ramus.height}{a numeric vector} }
}
\source{
Andrews and Herzberg (1985) Chapter 53.
}
\description{
Sex and species of an specimens of kangaroo.
}
\examples{

data(kanga)
## maybe str(kanga) ; plot(kanga) ...

}
\references{
Andrews, D. F. and Herzberg, A. M. (1985). Data.
Springer-Verlag, New York.
}
\keyword{datasets}
