% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{lawn}
\alias{lawn}
\title{Cut-off times of lawnmowers}
\format{
A data frame with 24 observations on the following 4 variables.
\describe{ \item{manufact}{Manufacturer - a factor with levels
\code{A} \code{B}} \item{machine}{Lawn mower - a factor with levels
\code{m1} \code{m2} \code{m3} \code{m4} \code{m5} \code{m6}}
\item{speed}{Speed of testing - a factor with levels \code{H}
\code{L}} \item{time}{cut-off time} }
}
\source{
Unknown.
}
\description{
Data on the cut-off times of lawnmowers was collected. 3 machines were
randomly selected from those produced by manufacturers A and B. Each machine
was tested twice at low speed and high speed.
}
\keyword{datasets}
