% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{leafblotch}
\alias{leafblotch}
\title{Leaf blotch on barley}
\format{
A data frame with 90 observations on the following 3 variables.
\describe{ \item{blotch}{proportion of the barley leaf affected by
blotch} \item{site}{the physical location - a factor with levels
\code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8}
\code{9}} \item{variety}{variety of barley - a factor with levels
\code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8}
\code{9} \code{10}} }
}
\source{
R. W. M. Wedderburn (1974) "Quasilikelihood functions, generalized
linear models and the Gauss-Newton method" Biometrika, 61, 439-447.
}
\description{
The data gives the proportion of leaf area affected by leaf blotch on 10
varieties of barley at 9 different sites.
}
\references{
P. McCullagh and J. Nelder (1989) "Generalized Linear Models"
Chapman and Hall, 2nd ed.
}
\keyword{datasets}
