% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{leafburn}
\alias{leafburn}
\title{Data on the burning time of samples of tobacco leaves}
\format{
A data frame with 30 observations on the following 4 variables.
\describe{ \item{nitrogen}{nitrogen content by percentage weight}
\item{chlorine}{chlorine content by percentage weight}
\item{potassium}{potassium content by percentage weight}
\item{burntime}{burn time in seconds} }
}
\source{
Steel, R. G. D. and Torrie, J. H. (1980), Principles and Procedures
of Statistics, Second Edition, New York: McGraw-Hill
}
\description{
Data on the burning time of samples of tobacco leaves
}
\keyword{datasets}
