% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway.R
\name{logit}
\alias{logit}
\title{Logit transformation}
\usage{
logit(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
log(x/(1-x))
}
\description{
Computes the logit transformation
}
\details{
x <=0 or >=1 will return NA
}
\examples{
logit(c(0.1,0.5,1.0,1.1))
#[1] -2.197225  0.000000        NA        NA

}
\seealso{
ilogit
}
\author{
Julian Faraway
}
