% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{mammalsleep}
\alias{mammalsleep}
\title{Sleep in Mammals: Ecological and Constitutional Correlates}
\format{
This data frame contains the following columns: \describe{
\item{body}{ body weight in kg } \item{brain}{ brain weight
in g } \item{nondream}{ slow wave ("nondreaming") sleep (hrs/day) }
\item{dream}{ paradoxical ("dreaming") sleep (hrs/day) }
\item{sleep}{ total sleep (hrs/day) (sum of slow wave and
paradoxical sleep) } \item{lifespan}{ maximum life span (years) }
\item{gestation}{ gestation time (days) } \item{predation}{
predation index (1-5) 1 = minimum (least likely to be preyed upon) to 5 =
maximum (most likely to be preyed upon) } \item{exposure}{ sleep
exposure index (1-5) 1 = least exposed (e.g. animal sleeps in a
well-protected den) 5 = most exposed } \item{danger}{ overall danger
index (1-5) (based on the above two indices and other information) 1 = least
danger (from other animals) 5 = most danger (from other animals) }}
}
\source{
"Sleep in Mammals: Ecological and Constitutional Correlates" by
Allison, T.  and Cicchetti, D. (1976), Science, November 12, vol. 194, pp.
732-734.
}
\description{
The \code{mammalsleep} data frame has 62 rows and 10 columns. Sleep in
Mammals: Ecological and Constitutional Correlates
}
\keyword{datasets}
