% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{manilius}
\alias{manilius}
\title{Mayer's 1750 data on the Manilius crater on the moon}
\format{
A data frame with 27 observations on the following 4 variables.
\describe{ \item{arc}{an angle known as h in Stigler's notation}
\item{sinang}{the sin(g-k) where g and k are two angles in Stigler}
\item{cosang}{the cos(g-k) where g and k are two angles in Stigler}
\item{group}{one of three groups determined by Mayer} }
}
\source{
Stigler, S. (1986) History of Statistics. Belknap Press, Harvard.
}
\description{
In 1750, Tobias Mayer collected data on various landmarks on the moon in
order to determine its orbit. The data involving the position of the
Manilius crater resulted in a least squares like problem. The example is
discussed in Steven Stigler's History of Statistics.
}
\details{
See Stigler for a detailed description.
}
\examples{

data(manilius)

}
\references{
Mayer, T. (1750) Abhandlung uber die Umwaltzung des Monds um
seine Axe und die scheinbare Bewegung der Mondsflecken published in the
Kosmographische Nachrichten und Sammlungen auf das Jahr 1748. 52-183
}
\keyword{datasets}
