% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway.R
\name{maxadjr}
\alias{maxadjr}
\title{Maximum Adjusted R-squared}
\usage{
maxadjr(l, best = 3)
}
\arguments{
\item{l}{A leaps object returned from leaps()}

\item{best}{An optional argument specify the number of models to be returned
taking the default value of 3}
}
\value{
A list of the best models
}
\description{
Displays the best models from a leaps object
}
\details{
Requires leaps package
}
\seealso{
leaps()
}
\author{
Julian Faraway
}
\keyword{regression}
