% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{melanoma}
\alias{melanoma}
\title{Melanoma by type and location}
\format{
A data frame with 12 observations on the following 3 variables.
\describe{ \item{count}{number of cases} \item{tumor}{type
of tumor - a factor with levels \code{freckle} \code{indeterminate}
\code{nodular} \code{superficial}} \item{site}{location of tumor on
the body - a factor with levels \code{extremity} \code{head} \code{trunk}} }
}
\source{
Dobson A. (2002) An introduction to generalized linear models,
Chapman Hall.
}
\description{
Data comes from a study of Malignant Melanoma involving 400 subjects.
}
\keyword{datasets}
