% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{nels88}
\alias{nels88}
\title{National Education Longitudinal Study of 1988}
\format{
A data frame with 260 observations on the following 5 variables.
\describe{ \item{sex}{a factor with levels \code{Female}
\code{Male}} \item{race}{a factor with levels \code{White}
\code{Asian} \code{Black} \code{Hispanic}} \item{ses}{a numeric
vector} \item{paredu}{a factor with levels \code{ba} \code{college}
\code{hs} \code{lesshs} \code{ma} \code{phd}} \item{math}{a numeric
vector} }
}
\source{
\url{http://www.icpsr.umich.edu/icpsrweb/ICPSR/series/107}
}
\description{
A subset of the National Education Longitudinal Study of 1988
}
\examples{

data(nels88)
## maybe str(nels88) ; plot(nels88) ...

}
\keyword{datasets}
