% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{nepali}
\alias{nepali}
\title{Nepali child heath study}
\format{
A data frame with 1000 observations on the following 9 variables.
\describe{ \item{id}{There is a six digit code for the child's ID: 2
digits for the panchayat number; 2 digits for the ward within panchayat; 1
digits for the household; 1 digit for child within household.}
\item{sex}{1 = male; 2 = female} \item{wt}{Child's weight
measured in kilograms} \item{ht}{Child's height measured in
centimeters} \item{mage}{Mother's age in years}
\item{lit}{Indicator of mother's literacy: 0 = no; 1 = yes}
\item{died}{The number of children the mother has had that died.}
\item{alive}{The number of children the mother has ever had born
alive} \item{age}{age of child} }
}
\source{
West KP, Jr., LeClerq SC, Shrestha SR, Wu LS, Pradhan EK, Khatry SK,
Katz J, Adhikari R, Sommer A.  Effects of vitamin A on growth of vitamin A
deficient children: field studies in Nepal. J Nutr 1997;10:1957-1965.
}
\description{
The data are a subset from public health study on Nepalese children.
}
\keyword{datasets}
