% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{oatvar}
\alias{oatvar}
\title{Yields of oat varieties planted in blocks}
\format{
The dataset contains the following variables \describe{
\item{yield}{ Yield in grams per 16ft row } \item{block}{
Blocks I to V } \item{variety}{ Variety 1 to 8} }
}
\source{
"Statistical Theory in Research" by R. Anderson and T. Bancroft,
McGraw Hill,1952
}
\description{
Data from an experiment to compare 8 varieties of oats. The growing area was
heterogeneous and so was grouped into 5 blocks. Each variety was sown once
within each block and the yield in grams per 16ft row was recorded.
}
\keyword{datasets}
