% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{odor}
\alias{odor}
\title{Odor of chemical by production settings}
\format{
\describe{ \item{odor}{ Odor score} \item{temp}{
Temperature coded as -1, 0 and 1} \item{gas}{ Gas/Liquid ratio coded
as -1, 0 and 1} \item{pack}{ Packing height coded as -1, 0 and 1} }
}
\source{
"Statistical Design and Analysis of Experiments" by P. John,
Macmillan, 1971
}
\description{
Data from an experiment to determine the effects of column temperature,
gas/liquid ratio and packing height in reducing unpleasant odor of chemical
product that was being sold for household use
}
\keyword{datasets}
