% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{orings}
\alias{orings}
\title{Spache Shuttle Challenger O-rings}
\format{
A data frame with 23 observations on the following 2 variables.
\describe{ \item{temp}{temperature at launch in degrees F}
\item{damage}{number of damage incidents out of 6 possible} }
}
\source{
Presidential Commission on the Space Shuttle Challenger Accident,
Vol. 1, 1986: 129-131.
}
\description{
The 1986 crash of the space shuttle Challenger was linked to failure of
O-ring seals in the rocket engines. Data was collected on the 23 previous
shuttle missions. The launch temperature on the day of the crash was 31F.
}
\references{
S. Dalal, E. Fowlkes and B. Hoadley (1989) "Risk Analysis of the
Space Shuttle: Pre-Challenger Prediction of Failure." Journal of the
American Statistical Association. 84: 945-957.
}
\keyword{datasets}
