% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{ozone}
\alias{ozone}
\title{Ozone in LA in 1976}
\format{
A data frame with 330 observations on the following 10 variables.
\describe{ \item{O3}{Ozone conc., ppm, at Sandbug AFB.}
\item{vh}{a numeric vector} \item{wind}{wind speed}
\item{humidity}{a numeric vector} \item{temp}{temperature}
\item{ibh}{inversion base height} \item{dpg}{Daggett
pressure gradient} \item{ibt}{a numeric vector}
\item{vis}{visibility} \item{doy}{day of the year} }
}
\source{
Breiman, L. and J. H. Friedman (1985). Estimating optimal
transformations for multiple regression and correlation. Journal of the
American Statistical Association 80, 580-598.
}
\description{
A study the relationship between atmospheric ozone concentration and
meteorology in the Los Angeles Basin in 1976.  A number of cases with
missing variables have been removed for simplicity.
}
\examples{

data(ozone)
## maybe str(ozone) ; plot(ozone) ...

}
\keyword{datasets}
