% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{parstum}
\alias{parstum}
\title{Marijuana and parent alcohol and drug use}
\format{
A data frame with 9 observations on the following 3 variables.
\describe{ \item{parent}{Number of parents using drugs or alcohol -
a factor with levels \code{Both} \code{Neither} \code{One}}
\item{student}{Student usage of marijuana - a factor with levels
\code{Never} \code{Occasional} \code{Regular}} \item{count}{the
number of cases} }
}
\source{
Ellis, Godfrey J. and Stone, Lorene H.  (1979) Marijuana Use in
College: "An Evaluation of a Modeling Explanation" Youth and Society 10,
323-34
}
\description{
445 college students were classified according to both frequency of
marijuana use and parental use of alcohol and psychoactive drugs.
}
\keyword{datasets}
