% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{peanut}
\alias{peanut}
\title{Carbon dioxide effects on peanut oil extraction}
\format{
This data frame contains the following columns: \describe{
\item{press}{ CO2 pressure - two levels low=0, high=1 }
\item{temp}{ CO2 temperature - two levels low=0, high=1 }
\item{moist}{ peanut moisture - two levels low=0, high=1 }
\item{flow}{ CO2 flow rate - two levels low=0, high=1 }
\item{size}{ peanut particle size - two levels low=0, high=1 }
\item{solubility}{ the amount of oil that could dissolve in the CO2
}}
}
\source{
Kilgo, M (1989) "An Application of Fractional Factorial Experimental
Designs" Quality Engineering, 1, 45-54
}
\description{
The \code{peanut} data frame has 16 rows and 6 columns. Carbon dioxide
effects on peanut oil extraction
}
\keyword{datasets}
