% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{phbirths}
\alias{phbirths}
\title{Birth weights in Philadelphia}
\format{
A data frame with 1115 observations on the following 5 variables.
\describe{ \item{black}{is the mother Black?}
\item{educ}{mother's years of education} \item{smoke}{does
the mother smoke during pregnancy?} \item{gestate}{gestational age
in weeks} \item{grams}{birth weight in grams} }
}
\source{
I. T. Elo, G. Rodriguez and H. Lee (2001). Racial and Neighborhood
Disparities in Birthweight in Philadelphia. Paper presented at the Annual
Meeting of the Population Association of America, Washington, DC 2001.
}
\description{
Data based on a 5% sample of all births occurring in Philadelphia in 1990.
}
\examples{

data(phbirths)
## maybe str(phbirths) ; plot(phbirths) ...

}
\keyword{datasets}
