% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{pima}
\alias{pima}
\title{Diabetes survey on Pima Indians}
\format{
The dataset contains the following variables \describe{
\item{pregnant}{ Number of times pregnant} \item{glucose}{
Plasma glucose concentration at 2 hours in an oral glucose tolerance test}
\item{diastolic}{ Diastolic blood pressure (mm Hg)}
\item{triceps}{ Triceps skin fold thickness (mm)}
\item{insulin}{ 2-Hour serum insulin (mu U/ml)} \item{bmi}{
Body mass index (weight in kg/(height in metres squared))}
\item{diabetes}{ Diabetes pedigree function} \item{age}{ Age
(years)} \item{test}{ test whether the patient shows signs of
diabetes (coded 0 if negative, 1 if positive)} }
}
\source{
The data may be obtained from UCI Repository of machine learning
databases at \url{http://archive.ics.uci.edu/ml/}
}
\description{
The National Institute of Diabetes and Digestive and Kidney Diseases
conducted a study on 768 adult female Pima Indians living near Phoenix.
}
\keyword{datasets}
