% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{pipeline}
\alias{pipeline}
\title{NIST data on ultrasonic measurements of defects in the Alaska pipeline}
\format{
A data frame with 107 observations on the following 3 variables.
\describe{ \item{Field}{measurement of depth of defect on site}
\item{Lab}{measurement of depth of defect in the lab} \item{Batch}{the batch
of measurements} }
}
\source{
Office of the Director of the Institute of Materials Research (now
the Materials Science and Engineering Laboratory) of NIST
}
\description{
Researchers at National Institutes of Standards and Technology (NIST)
collected data on ultrasonic measurements of the depths of defects in the
Alaska pipeline in the field. The depth of the defects were then remeasured
in the laboratory. These measurements were performed in six different
batches. The laboratory measurements are more accurate than the in-field
measurements, but more time consuming and expensive.
}
\keyword{datasets}
