% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{pneumo}
\alias{pneumo}
\title{Pneumonoconiosis in coal miners}
\format{
A data frame with 24 observations on the following 3 variables.
\describe{ \item{Freq}{number of miners}
\item{status}{pneumoconiosis status - a factor with levels
\code{mild} \code{normal} \code{severe}} \item{year}{number of years
service (midpoint of interval)} }
}
\source{
M. Aitkin and D. Anderson and B. Francis and J. Hinde (1989)
"Statistical Modelling in GLIM" Oxford University Press.
}
\description{
The data for this example contains the number of coal miners classified by
radiological examination into one of three categories of
pneumonoultramicroscopicosilicovolcanoconiosis (known as pneumonoconiosis
for short) and by number of years spent working at the coal face divided
into eight categories.
}
\keyword{datasets}
