% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{potuse}
\alias{potuse}
\title{Marijuana usage by youth}
\format{
A data frame with 486 observations on the following 7 variables.
\describe{ \item{sex}{1=Male, 2=Female}
\item{year.76}{1=never used, 2=used no more than once a month,
3=used more than once a month in 1976} \item{year.77}{1=never used,
2=used no more than once a month, 3=used more than once a month in 1977}
\item{year.78}{1=never used, 2=used no more than once a month,
3=used more than once a month in 1978} \item{year.79}{1=never used,
2=used no more than once a month, 3=used more than once a month in 1979}
\item{year.80}{1=never used, 2=used no more than once a month,
3=used more than once a month in 1980} \item{count}{Number of cases
in this category} }
}
\source{
ICPSR, University of Michigan
}
\description{
The National Youth Survey collected a sample of 11 to 17 year olds - 117
boys and 120 girls - asking questions about marijuana usage.
}
\references{
Lang J., McDonald, J and Smith P. (1999) "Association-Marginal
Modeling of Mutlivariate Categorical Responses: A Maximum Likelihood
Approach" JASA 94, 1161-
}
\keyword{datasets}
