% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{prostate}
\alias{prostate}
\title{Prostate cancer surgery}
\format{
This data frame contains the following columns: \describe{

\item{lcavol}{ log(cancer volume) } \item{lweight}{
log(prostate weight) }

\item{age}{ age } \item{lbph}{ log(benign prostatic
hyperplasia amount) } \item{svi}{ seminal vesicle invasion }
\item{lcp}{ log(capsular penetration) } \item{gleason}{
Gleason score } \item{pgg45}{ percentage Gleason scores 4 or 5 }
\item{lpsa}{ log(prostate specific antigen) } }
}
\source{
Stamey, T.A., Kabalin, J.N., McNeal, J.E., Johnstone, I.M., Freiha,
 F., Redwine, E.A. and Yang, N. (1989). Prostate specific antigen in
 the diagnosis and treatment of adenocarcinoma of the prostate:
 II. radical prostatectomy treated patients. Journal of Urology
 141(5):1076–1083. 
}
\description{
The \code{prostate} data frame has 97 rows and 9 columns. A study on 97 men
with prostate cancer who were due to receive a radical prostatectomy.
}
\keyword{datasets}
