% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway.R
\name{prplot}
\alias{prplot}
\title{Partial Residual Plot}
\usage{
prplot(g, i)
}
\arguments{
\item{g}{An object returned from lm()}

\item{i}{index of predictor}
}
\value{
none
}
\description{
Makes a Partial Residual plot
}
\examples{

data(stackloss)
lmod <- lm(stack.loss ~ .,stackloss)
prplot(lmod,1)

}
\author{
Julian Faraway
}
\keyword{regression}
