% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{psid}
\alias{psid}
\title{Panel Study of Income Dynamics subset}
\format{
A data frame with 1661 observations on the following 6 variables.
\describe{ \item{age}{age in 1968} \item{educ}{years of
education} \item{sex}{sex of individual, \code{F} or \code{M}}
\item{income}{annual income in dollars} \item{year}{calendar
year} \item{person}{ID number for individual} }
}
\source{
Martha S. Hill, The Panel Study of Income Dynamics: A User's Guide,
Sage Publications, 1992,Newbury Park, CA.
}
\description{
The Panel Study of Income Dynamics (PSID), begun in 1968, is a longitudinal
study of a representative sample of U.S. individuals.  The study is
conducted at the Survey Research Center, Institute for Social Research,
University of Michigan and is still continuing. The data represents a small
subset of the total data.
}
\keyword{datasets}
