% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{pulp}
\alias{pulp}
\title{Brightness of paper pulp depending on shift operator}
\format{
This data frame contains the following columns: \describe{
\item{bright}{ Brightness of the pulp as measured by a reflectance
meter } \item{operator}{ Shift operator a-d }}
}
\source{
"Statistical techniques applied to production situations" F. Sheldon
(1960) Industrial and Engineering Chemistry, 52, 507-509
}
\description{
The \code{pulp} data frame has 20 rows and 2 columns. Data comes from an
experiment to test the paper brightness depending on a shift operator.
}
\keyword{datasets}
