% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{punting}
\alias{punting}
\title{Leg strength and punting}
\format{
A data frame with 13 observations on the following 7 variables.
\describe{ \item{Distance}{average distance over 10 punts}
\item{Hang}{hang time} \item{RStr}{right leg strength in
pounds} \item{LStr}{left leg strength in pounds}
\item{RFlex}{right hamstring muscle flexibility in degrees}
\item{LFlex}{left hamstring muscle flexibility in degrees}
\item{OStr}{overall leg strength in foot pounds} }
}
\source{
Unknown
}
\description{
Investigators studied physical characteristics and ability in 13 (American)
football punters. Each volunteer punted a football ten times. The
investigators recorded the average distance for the ten punts, in feet.
}
\examples{

data(punting)
## maybe str(punting) ; plot(punting) ...

}
\keyword{datasets}
