% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{pvc}
\alias{pvc}
\title{Production of PVC by operator and resin railcar}
\format{
Dataset contains the following variables \describe{
\item{psize}{ Particle size} \item{operator}{ Operator
number 1, 2 or 3} \item{resin}{ Resin railcar 1-8} }
}
\source{
R. Morris and E. Watson (1998) "A comparison of the techniques used
to evaluate the measurement process" Quality Engineering, 11, 213-219
}
\description{
Data from an experiment to study factors affecting the production of the
plastic PVC, 3 operators used 8 different devices called resin railcars to
produce PVC. For each of the 24 combinations, two samples were produced.
}
\keyword{datasets}
