% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{pyrimidines}
\alias{pyrimidines}
\title{Activity in pyrimidines}
\format{
A data frame with 74 observations on the following 27 variables.
\describe{ \item{p1.polar}{measured on a [0,1] scale}
\item{p1.size}{measured on a [0,1] scale}
\item{p1.flex}{measured on a [0,1] scale}
\item{p1.h.doner}{measured on a [0,1] scale}
\item{p1.h.acceptor}{measured on a [0,1] scale}
\item{p1.pi.doner}{measured on a [0,1] scale}
\item{p1.pi.acceptor}{measured on a [0,1] scale}
\item{p1.polarisable}{measured on a [0,1] scale}
\item{p1.sigma}{measured on a [0,1] scale}
\item{p2.polar}{measured on a [0,1] scale}
\item{p2.size}{measured on a [0,1] scale}
\item{p2.flex}{measured on a [0,1] scale}
\item{p2.h.doner}{measured on a [0,1] scale}
\item{p2.h.acceptor}{measured on a [0,1] scale}
\item{p2.pi.doner}{measured on a [0,1] scale}
\item{p2.pi.acceptor}{measured on a [0,1] scale}
\item{p2.polarisable}{measured on a [0,1] scale}
\item{p2.sigma}{measured on a [0,1] scale}
\item{p3.polar}{measured on a [0,1] scale}
\item{p3.size}{measured on a [0,1] scale}
\item{p3.flex}{measured on a [0,1] scale}
\item{p3.h.doner}{measured on a [0,1] scale}
\item{p3.h.acceptor}{measured on a [0,1] scale}
\item{p3.pi.doner}{measured on a [0,1] scale}
\item{p3.polarisable}{measured on a [0,1] scale}
\item{p3.sigma}{measured on a [0,1] scale}
\item{activity}{log 1/Ki, where Ki is the inhibition constant as
experimentally assayed, scaled to [0,1]} }
}
\source{
Jonathan D. Hirst, Ross D. King, Michael J. E. Sternberg (1994)
Quantitative structure-activity relationships by neural networks and
inductive logic programming. I. The inhibition of dihydrofolate reductase by
pyrimidines \doi{10.1007/BF00125375}
}
\description{
Structural information on 74 2,4-diamino- 5-(substituted benzyl) pyrimidines
used as inhibitors of DHFR in E. coli. There are 3 positions where chemical
activity occurs and 9 attributes per position leading to 27 total
predictors. One predictor had no variability and was removed from the data
set. 26 chemical properties of 74 compounds and an activity level
}
\examples{

data(pyrimidines)
## maybe str(pyrimidines) ; plot(pyrimidines) ...

}
\keyword{datasets}
