% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway.R
\name{qqnorml}
\alias{qqnorml}
\title{Labeled QQ plot}
\usage{
qqnorml(
  y,
  main = "Normal Q-Q Plot",
  xlab = "Theoretical Quantiles",
  ylab = "Sample Quantiles",
  ...
)
}
\arguments{
\item{y}{A numeric vector}

\item{main}{main label}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{arguments passed to plot()}
}
\value{
none
}
\description{
Makes a labeled QQ plot
}
\examples{

qqnorml(rnorm(16))

}
\seealso{
qqnorm
}
\author{
Julian Faraway
}
