% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{rabbit}
\alias{rabbit}
\title{Rabbit weight gain by diet and litter}
\format{
The variables in the dataset were \describe{ \item{treat}{
Diet a through f}

\item{gain}{ Weight gain}

\item{block}{ Block (10 litters)} }
}
\source{
"Experimental Design and Analysis" by M. Lentner and T. Bishop,
Valley Book Company, 1986
}
\description{
A nutritionist studied the effects of six diets, on weight gain of domestic
rabbits.  From past experience with sizes of litters, it was felt that only
3 uniform rabbits could be selected from each available litter. There were
ten litters available forming blocks of size three.
}
\keyword{datasets}
