% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{ratdrink}
\alias{ratdrink}
\title{Rat growth weights affected by additives}
\format{
A data frame with 135 observations on the following 4 variables.
\describe{ \item{wt}{Weight of the rat} \item{weeks}{Week of
the study from 0 to 4} \item{subject}{the rat code number}
\item{treat}{treatment applied to the rat drinking water - a factor
with levels \code{control} \code{thiouracil} \code{thyroxine}} }
}
\source{
Unknown
}
\description{
The data consist of 5 weekly measurements of body weight for 27 rats. The
first 10 rats are on a control treatment while 7 rats have thyroxine added
to their drinking water. 10 Rats have thiouracil added to their water.
}
\keyword{datasets}
