\name{rats}
\alias{rats}
\docType{data}
\title{Effect of toxic agents on rats}
\description{
  An experiment was conducted as part of an investigation to combat the
  effects of certain toxic agents. 
}
\usage{data(rats)}
\format{
  A data frame with 48 observations on the following 3 variables.
  \describe{
    \item{\code{time}}{survival time in tens of hours}
    \item{\code{poison}}{the poison type - a factor with levels \code{I} \code{II} \code{III}}
    \item{\code{treat}}{the treatment - a factor with levels \code{A} \code{B} \code{C} \code{D}}
  }
}
\source{
  Box G and Cox D (1964) "An analysis of transformations"
  J. Roy. Stat. Soc. Series B. \bold{26} 211.
}
\keyword{datasets}
