% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{rats}
\alias{rats}
\title{Effect of toxic agents on rats}
\format{
A data frame with 48 observations on the following 3 variables.
\describe{ \item{time}{survival time in tens of hours}
\item{poison}{the poison type - a factor with levels \code{I}
\code{II} \code{III}} \item{treat}{the treatment - a factor with
levels \code{A} \code{B} \code{C} \code{D}} }
}
\source{
Box G and Cox D (1964) "An analysis of transformations" J. Roy.
Stat. Soc. Series B. \bold{26} 211.
}
\description{
An experiment was conducted as part of an investigation to combat the
effects of certain toxic agents.
}
\keyword{datasets}
