% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{resceram}
\alias{resceram}
\title{Shape and plate effects on current noise in resistors}
\format{
This data frame contains the following columns: \describe{
\item{noise}{ current noise } \item{shape}{ the geometrical
shape of the resistor, A, B, C or D } \item{plate}{ the ceramic
plate on which the resistor was mounted. Only three resistors will fit on
one plate. }}
}
\source{
Natrella, M (1963) "Experimental Statistics" National Bureau of
Standards Handbook 91, Gaithersburg MD.
}
\description{
The \code{resceram} data frame has 12 rows and 3 columns. Shape and plate
effects on current noise in resistors
}
\keyword{datasets}
