% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{salmonella}
\alias{salmonella}
\title{Salmonella reverse mutagenicity assay}
\format{
A data frame with 18 observations on the following 2 variables.
\describe{ \item{colonies}{numbers of revertant colonies of TA98 Salmonella}
\item{dose}{dose level of quinoline} }
}
\source{
Breslow N.E. (1984), Extra-Poisson Variation in Log-linear Models,
ApplStat, pp. 38-44.
}
\description{
The data was collected in a salmonella reverse mutagenicity assay where the
numbers of revertant colonies of TA98 Salmonella observed on each of three
replicate plates for different doses of quinoline
}
\keyword{datasets}
