% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{sat}
\alias{sat}
\title{School expenditure and test scores from USA in 1994-95}
\format{
This data frame contains the following columns: \describe{
\item{expend}{ Current expenditure per pupil in average daily
attendance in public elementary and secondary schools, 1994-95 (in thousands
of dollars) } \item{ratio}{ Average pupil/teacher ratio in public
elementary and secondary schools, Fall 1994 } \item{salary}{
Estimated average annual salary of teachers in public elementary and
secondary schools, 1994-95 (in thousands of dollars) }
\item{takers}{ Percentage of all eligible students taking the SAT,
1994-95 } \item{verbal}{ Average verbal SAT score, 1994-95 }
\item{math}{ Average math SAT score, 1994-95 } \item{total}{
Average total score on the SAT, 1994-95 } }
}
\source{
"Getting What You Pay For: The Debate Over Equity in Public School
Expenditures" D. Guber, Journal of Statistics Education, 1999
}
\description{
The \code{sat} data frame has 50 rows and 7 columns.  Data were collected to
study the relationship between expenditures on public education and test
results.
}
\keyword{datasets}
