% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{savings}
\alias{savings}
\title{Savings rates in 50 countries}
\format{
This data frame contains the following columns: \describe{

\item{sr}{savings rate - personal saving divided by disposable
income}

\item{pop15}{percent population under age of 15}

\item{pop75}{percent population over age of 75}

\item{dpi}{per-capita disposable income in dollars}

\item{ddpi}{percent growth rate of dpi} }
}
\source{
Belsley, D., Kuh. E. and Welsch, R. (1980) "Regression Diagnostics"
Wiley.
}
\description{
The \code{savings} data frame has 50 rows and 5 columns.  The data is
averaged over the period 1960-1970.
}
\details{
Now also appears as \code{LifeCycleSavings} in the \code{datasets} package
}
\seealso{
LifeCycleSavings
}
\keyword{datasets}
