% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{seatpos}
\alias{seatpos}
\title{Car seat position depending driver size}
\format{
The dataset contains the following variables \describe{
\item{Age}{ Age in years} \item{Weight}{ Weight in lbs}
\item{HtShoes}{ Height in shoes in cm} \item{Ht}{ Height
bare foot in cm} \item{Seated}{ Seated height in cm}
\item{Arm}{ lower arm length in cm} \item{Thigh}{ Thigh
length in cm} \item{Leg}{ Lower leg length in cm}
\item{hipcenter}{ horizontal distance of the midpoint of the hips
from a fixed location in the car in mm} }
}
\source{
"Linear Models in R" by Julian Faraway, CRC Press, 2004
}
\description{
Car drivers like to adjust the seat position for their own comfort. Car
designers would find it helpful to know where different drivers will
position the seat depending on their size and age. Researchers at the
HuMoSim laboratory at the University of Michigan collected data on 38
drivers.
}
\keyword{datasets}
