% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{snail}
\alias{snail}
\title{Snail production}
\format{
A data frame with 24 observations on the following 3 variables.
\describe{ \item{water}{percentage water content}
\item{temp}{temperature in C} \item{humid}{relative
humidity} }
}
\source{
Unknown
}
\description{
A study was conducted to optimize snail production for consumption. The
percentage water content of the tissues of snails grown under three
different levels of relative humidity and two different temperatures was
recorded. For each combination, 4 snails were observed.
}
\examples{

data(snail)
## maybe str(snail) ; plot(snail) ...

}
\keyword{datasets}
