% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{solder}
\alias{solder}
\title{Solder skips in printing circuit boards}
\format{
A data frame with 900 observations on the following 6 variables.
\describe{ \item{Opening}{a factor with levels \code{L} \code{M}
\code{S}} \item{Solder}{a factor with levels \code{Thick}
\code{Thin}} \item{Mask}{a factor with levels \code{A1.5} \code{A3}
\code{A6} \code{B3} \code{B6}} \item{PadType}{a factor with levels
\code{D4} \code{D6} \code{D7} \code{L4} \code{L6} \code{L7} \code{L8}
\code{L9} \code{W4} \code{W9}} \item{Panel}{a numeric vector}
\item{skips}{count of how many solder skips appeared to a visual
inspection} }
}
\source{
Comizzoli, R. B., J. M. Landwehr, and J. D. Sinclair (1990). Robust
materials and processes: Key to reliability. AT&T Technical Journal 69(6),
113-128.
}
\description{
ATT ran an experiment varying five factors relevant to a wave-soldering
procedure for mounting components on printed circuit boards.  The response
variable, skips, is a count of how many solder skips appeared to a visual
inspection.
}
\examples{

data(solder)
## maybe str(solder) ; plot(solder) ...

}
\keyword{datasets}
