% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{soybean}
\alias{soybean}
\title{Germination failures for soybean seeds}
\format{
A data frame with 25 observations on the following 3 variables.
\describe{ \item{variety}{the variety - a factor with levels
\code{arasan} \code{check} \code{fermate} \code{semesan} \code{spergon}}
\item{replicate}{the plot - a factor with levels \code{1} \code{2}
\code{3} \code{4} \code{5}} \item{failure}{the number of failures
out of 100 planted seeds} }
}
\source{
Snedecor G. and Cochran W. (1967) Statistical Methods (6th Ed) Iowa
State University Press
}
\description{
An experiment was conducted to compare the germination rates of the five
varieties of soybean. Five plots were available.
}
\keyword{datasets}
