% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{speedo}
\alias{speedo}
\title{Speedometer cable shrinkage}
\format{
The dataset contains the following variables: (variables a-o are 2
level factors, coded "+" and "-" where "+" indicates a higher value where
appropriate) \describe{ \item{a}{ liner outer diameter}
\item{b}{ liner die} \item{c}{ liner material}
\item{d}{ liner line speed} \item{e}{ wire braid type}
\item{f}{ braiding tension} \item{g}{ wire diameter}
\item{h}{ liner tension} \item{i}{ liner temperature}
\item{j}{ coating material} \item{k}{ coating die type}
\item{l}{ melt temperature} \item{m}{ screen pack}
\item{n}{ cooling method} \item{o}{ line speed}
\item{y}{ percentage shrinkage per specimen} }
}
\source{
G. P. Box and S. Bisgaard and C. Fung (1988) "An explanation and
critque of Taguchi's contributions to quality engineering", Quality and
reliability engineering international, 4, 123-131
}
\description{
Speedometer cables can be noisy because of shrinkage in the plastic casing
material. An experiment was conducted to find out what caused shrinkage by
screening a large number of factors.  The engineers started with 15
different factors.
}
\keyword{datasets}
