% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{star}
\alias{star}
\title{Star temperatures and light intensites}
\format{
A data frame with 47 observations on the following 3 variables.
\describe{ \item{index}{a numeric vector}
\item{temp}{temperature} \item{light}{light intensity} }
}
\source{
Rousseeuw, P. and A. Leroy (1987). Robust Regression and Outlier
Detection. New York: Wiley.
}
\description{
Data on the log of the surface temperature and the log of the light
intensity of 47 stars in the star cluster CYG OB1, which is in the direction
of Cygnus,
}
\examples{

data(star)
## maybe str(star) ; plot(star) ...

}
\keyword{datasets}
