\name{stat500}
\alias{stat500}
\docType{data}
\title{
Marks in a statistics class
}
\description{
Marks from Statistics 500 one year at the University of Michigan
}
\usage{data(stat500)}
\format{
  A data frame with 55 observations on the following 4 variables.
  \describe{
    \item{\code{midterm}}{a numeric vector}
    \item{\code{final}}{a numeric vector}
    \item{\code{hw}}{a numeric vector}
    \item{\code{total}}{a numeric vector}
  }
}
\source{
Julian Faraway 
}
\examples{
data(stat500)
## maybe str(stat500) ; plot(stat500) ...
}
\keyword{datasets}
