% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{stat500}
\alias{stat500}
\title{Marks in a statistics class}
\format{
A data frame with 55 observations on the following 4 variables.
\describe{ \item{midterm}{a numeric vector} \item{final}{a
numeric vector} \item{hw}{a numeric vector} \item{total}{a
numeric vector} }
}
\source{
Julian Faraway
}
\description{
Marks from Statistics 500 one year at the University of Michigan
}
\examples{

data(stat500)
## maybe str(stat500) ; plot(stat500) ...

}
\keyword{datasets}
