% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{stepping}
\alias{stepping}
\title{Stepping and effect on heart rate}
\format{
A data frame with 30 observations on the following 6 variables.
\describe{ \item{Order}{running order within the experiment}
\item{Block}{Experimenter used} \item{Height}{0 if step at
the low (5.75in) height, 1 if at the high (11.5in) height}
\item{Frequency}{the rate of stepping. 0 if slow (14 steps/min), 1
if medium (21 steps/min), 2 if high (28 steps/min)}
\item{RestHR}{the resting heart rate of the subject before a trial,
in beats per minute} \item{HR}{ the final heart rate of the subject
after a trial, in beats per minute } }
}
\source{
Unknown
}
\description{
An experiment was conducted to explore the nature of the relationship
between a person's heart rate and the frequency at which that person stepped
up and down on steps of various heights.
}
\examples{

data(stepping)
## maybe str(stepping) ; plot(stepping) ...

}
\keyword{datasets}
