% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{strongx}
\alias{strongx}
\title{Strong interaction experiment data}
\format{
\describe{Dataframe with 10 cases
\item{momentum}{}
\item{energy}{inverse total energy}
\item{crossx}{Scattering cross-section/sec}
\item{sd}{standard deviation}}
}
\source{
Weisberg, H., Beier, H., Brody, H., Patton, R., Raychaudhari, K., Takeda, H., Thern, R. and Van Berg, R. (1978). s-dependence of proton fragmentation by hadrons. II. Incident laboratory momenta, 30–250 GeV/c. Physics Review D, 17, 2875–2887.
}
\description{
Example Dataset from "Practical Regression and Anova"
}
\references{
Weisberg, S. (2014). Applied Linear Regression, 4th edition. Hoboken NJ: Wiley.
}
\keyword{datasets}
