% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway.R
\docType{methods}
\name{sumary}
\alias{sumary}
\alias{sumary,lm-method}
\alias{sumary,glm-method}
\alias{sumary,merMod-method}
\title{Abbreviated Regression Summary}
\arguments{
\item{object}{An lm, glm or mer object returned from lm(), glm() or lmer()
respectively}

\item{...}{further arguments passed to or from other methods.}
}
\value{
returns the same as \code{summary()}
}
\description{
Generic summaries for lm, glm and mer objects
}
\details{
This generic function provides an abbreviated regression output containing
the more useful information. Users wanting to see more are advised to use
\code{summary()}
}
\examples{

data(stackloss)
object <- lm(stack.loss ~ .,stackloss)
sumary(object)

}
\references{
This function is adapted from the \code{display()} function in
the \code{arm} package
}
\seealso{
\code{\link[base]{summary}}, \code{\link[stats]{lm}},
\code{\link[stats]{glm}}, \code{\link[lme4]{lmer}}
}
\author{
Julian Faraway
}
\keyword{regression}
