% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{teengamb}
\alias{teengamb}
\title{Study of teenage gambling in Britain}
\format{
This data frame contains the following columns: \describe{
\item{sex}{ 0=male, 1=female } \item{status}{ Socioeconomic
status score based on parents' occupation } \item{income}{ in pounds
per week } \item{verbal}{ verbal score in words out of 12 correctly
defined } \item{gamble}{ expenditure on gambling in pounds per year
}}
}
\source{
Ide-Smith & Lea, 1988, Journal of Gambling Behavior, 4, 110-118
}
\description{
The \code{teengamb} data frame has 47 rows and 5 columns. A survey was
conducted to study teenage gambling in Britain.
}
\keyword{datasets}
