% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{truck}
\alias{truck}
\title{Truck leaf spring experiment}
\format{
A data frame with 48 observations on the following 6 variables.
\describe{ \item{B}{furnace temperature - a factor with levels
\code{+} \code{-}} \item{C}{heating time - a factor with levels
\code{+} \code{-}} \item{D}{transfer time - a factor with levels
\code{+} \code{-}} \item{E}{hold-down time - a factor with levels
\code{+} \code{-}} \item{O}{quench oil temperature - a factor with
levels \code{+} \code{-}} \item{height}{leaf spring free height in
inches} }
}
\source{
J. J. Pignatiello and J. S. Ramberg (1985) Contribution to
discussion of offline quality control, parameter design and the Taguchi
method, Journal of Quality Technology, \bold{17} 198-206.
}
\description{
Data on an experiment concerning the production of leaf springs for trucks.
A \eqn{2^{5-1}} fractional factorial experiment with 3 replicates was
carried out with objective of recommending production settings to achieve a
free height as close as possible to 8 inches.
}
\references{
P. McCullagh and J. Nelder (1989) "Generalized Linear Models"
Chapman and Hall, 2nd ed.
}
\keyword{datasets}
