% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{tvdoctor}
\alias{tvdoctor}
\title{Life, TVs and Doctors}
\format{
A data frame with 38 observations on the following 3 variables.
\describe{ \item{life}{Life expectancy in years}
\item{tv}{Number of people per television set}
\item{doctor}{Number of people per doctor} }
}
\source{
Unknown, data for illustration purposes only
}
\description{
Life expectancy, doctors and televisions collected on 38 countries in 1993
}
\examples{

data(tvdoctor)
## maybe str(tvdoctor) ; plot(tvdoctor) ...

}
\keyword{datasets}
