% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{twins}
\alias{twins}
\title{Twin IQs from Burt}
\format{
A dataframe with the following variables: \describe{
\item{Foster}{IQ of the fostered child}
\item{Biological}{IQ of the biological child}
\item{Social}{social class of natural parents}}
}
\source{
Burt, C. (1966). The genetic estimation of differences in intelligence: 
A study of monozygotic twins reared together and apart. Br. J. Psych., 57, 147-153.
}
\description{
Study of IQ in twins reared apart
}
\references{
Weisberg, S. (2014). Applied Linear Regression, 4th edition. Hoboken NJ: Wiley.
}
\keyword{datasets}
