\name{wheat}
\alias{wheat}
\docType{data}
\title{
Insect damage to wheat by variety
}
\description{
Insect damage to wheat by variety
}
\usage{data(wheat)}
\format{
  A data frame with 13 observations on the following 2 variables.
  \describe{
    \item{\code{damage}}{a numeric vector}
    \item{\code{variety}}{a factor with levels \code{A} \code{B} \code{C} \code{D}}
  }
}
\source{
Unknown
}
\examples{
data(wheat)
## maybe str(wheat) ; plot(wheat) ...
}
\keyword{datasets}
