% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{wheat}
\alias{wheat}
\title{Insect damage to wheat by variety}
\format{
A data frame with 13 observations on the following 2 variables.
\describe{ \item{damage}{a numeric vector} \item{variety}{a
factor with levels \code{A} \code{B} \code{C} \code{D}} }
}
\source{
Unknown
}
\description{
Insect damage to wheat by variety
}
\examples{

data(wheat)
## maybe str(wheat) ; plot(wheat) ...

}
\keyword{datasets}
