% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faraway-package.R
\docType{data}
\name{worldcup}
\alias{worldcup}
\title{Data on players from the 2010 World Cup}
\format{
A data frame with 595 observations on the following 7 variables.
\describe{ \item{Team}{Country} \item{Position}{a factor
with levels \code{Defender} \code{Forward} \code{Goalkeeper}
\code{Midfielder}} \item{Time}{Time played in minutes}
\item{Shots}{Number of shots attempted} \item{Passes}{Number
of passes made} \item{Tackles}{Number of tackles made}
\item{Saves}{Number of saves made} }
}
\source{
Lost
}
\description{
Data on players from the 2010 World Cup
}
\details{
None
}
\examples{

data(worldcup)
## maybe str(worldcup) ; plot(worldcup) ...

}
\keyword{datasets}
